Random data.table generator

Generate random data.tables.

Usage

any_data.table(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_data.table() |> show_example()
#>          ...1      ...2  ...3 ...4            ...5
#> 1: 2829-11-07 <list[2]> FALSE TRUE 08:18:33.575708
any_data.table(rows = 10L) |> show_example()
#>          ...1       ...2               ...3  ...4       ...5  ...6      ...7
#>  1:    Mpj.f+ 1223-04-08               3444     0 Xx.mk+\\9| -8473 <list[2]>
#>  2:    -ms6f] 1362-10-11               TRUE  4159      Pb|~" -7441 <list[2]>
#>  3: c5d m@BaL 2018-12-15               1508  -670       oSU4 -8811 <list[2]>
#>  4:       .<; 1734-11-03           D)/d7~WK -7206    kFGk_Vi -1569 <list[2]>
#>  5:     V'mR& 2681-06-07 337-04-13 01:35:34 -5707          _  3940 <list[2]>
#>  6:       'qF 2400-04-17 747-11-24 06:31:14  3021    ]MRJ5[U  -160 <list[2]>
#>  7:       )hS 1792-01-23    08:52:06.263885  6417          q   945 <list[2]>
#>  8:   HOC<[0S 2832-09-02    14:01:39.045926  2968         gz -7587 <list[2]>
#>  9: vUJo>0E$P 2482-04-07              -7491 -8219         U9     0 <list[2]>
#> 10:     E/N-Q 1608-07-11          ,#xMM5lo,  7078         xx -9304 <list[2]>
#>         ...8  ...9 ...10
#>  1:   Ja ?}d FALSE  4328
#>  2:       .W  TRUE  4586
#>  3:        m  TRUE  2523
#>  4:    K)(KP FALSE  2808
#>  5:        / FALSE -5019
#>  6:       w/  TRUE  -704
#>  7:  XB-\\Z& FALSE     0
#>  8: ~G|@d7+m FALSE -3808
#>  9: !Pk\\$?E  TRUE     0
#> 10:  &?p2Lo( FALSE -5405
any_data.table(cols = 5L, any_na = TRUE) |> show_example()
#>          ...1      ...2            ...3  ...4      ...5
#> 1: 1297-01-30 <list[2]> 08:38:19.855402     0 <list[2]>
#> 2: 1908-09-08 <list[2]> 21:28:51.887439    NA <list[2]>
#> 3: 1790-02-03 <list[2]> 16:56:50.399437  7726 <list[2]>
#> 4: 2958-08-03 <list[2]> 12:30:51.198699    NA <list[2]>
#> 5: 1022-01-07 <list[2]> 07:38:21.604899  2245 <list[2]>
#> 6: 1001-01-02 <list[2]> 05:42:17.160217  4180 <list[2]>
#> 7:       <NA> <list[2]> 23:22:50.736630 -8116 <list[2]>
#> 8: 1458-01-04 <list[2]>              NA  7599 <list[2]>