Any data frame generator

Generate random data frames.

Usage

any_data_frame(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_data_frame() |> show_example()
#>                    ...1                ...2                ...3       ...4
#> 1 -5568521352, -8519171 1740-07-04 07:27:22 1, 3, 6, 4, 2, 2, 5 -380635535
#> 2 -5568521352, -8519171 1740-07-04 07:27:22 1, 3, 6, 4, 2, 2, 5 -380635535
#> 3 -5568521352, -8519171 1740-07-04 07:27:22 1, 3, 6, 4, 2, 2, 5 -380635535
#> 4 -5568521352, -8519171 1740-07-04 07:27:22 1, 3, 6, 4, 2, 2, 5 -380635535
#> 5 -5568521352, -8519171 1740-07-04 07:27:22 1, 3, 6, 4, 2, 2, 5 -380635535
#> 6 -5568521352, -8519171 1740-07-04 07:27:22 1, 3, 6, 4, 2, 2, 5 -380635535
#> 7 -5568521352, -8519171 1740-07-04 07:27:22 1, 3, 6, 4, 2, 2, 5 -380635535
#>                                                                ...5 ...6
#> 1 1256506, -4629536, 9435273, -6668261, -8076982, 5789283, -9004672 TRUE
#> 2 1256506, -4629536, 9435273, -6668261, -8076982, 5789283, -9004672 TRUE
#> 3 1256506, -4629536, 9435273, -6668261, -8076982, 5789283, -9004672 TRUE
#> 4 1256506, -4629536, 9435273, -6668261, -8076982, 5789283, -9004672 TRUE
#> 5 1256506, -4629536, 9435273, -6668261, -8076982, 5789283, -9004672 TRUE
#> 6 1256506, -4629536, 9435273, -6668261, -8076982, 5789283, -9004672 TRUE
#> 7 1256506, -4629536, 9435273, -6668261, -8076982, 5789283, -9004672 TRUE
#>                                                           ...7 ...8       ...9
#> 1 -24096045787, -57369529688, -50380008698, 714969847, 1, 3, 2    1 -793389533
#> 2 -24096045787, -57369529688, -50380008698, 714969847, 1, 3, 2    1 -793389533
#> 3 -24096045787, -57369529688, -50380008698, 714969847, 1, 3, 2    1 -793389533
#> 4 -24096045787, -57369529688, -50380008698, 714969847, 1, 3, 2    1 -793389533
#> 5 -24096045787, -57369529688, -50380008698, 714969847, 1, 3, 2    1 -793389533
#> 6 -24096045787, -57369529688, -50380008698, 714969847, 1, 3, 2    1 -793389533
#> 7 -24096045787, -57369529688, -50380008698, 714969847, 1, 3, 2    1 -793389533
any_data_frame(rows = 10L) |> show_example()
#>   ...1                                                  ...2      ...3
#> 1    P 0, 0, 0, 342754495, 119409368, -735727453, -968714725 518334522
#> 2    P 0, 0, 0, 342754495, 119409368, -735727453, -968714725 518334522
#> 3    P 0, 0, 0, 342754495, 119409368, -735727453, -968714725 518334522
#> 4    P 0, 0, 0, 342754495, 119409368, -735727453, -968714725 518334522
#> 5    P 0, 0, 0, 342754495, 119409368, -735727453, -968714725 518334522
#> 6    P 0, 0, 0, 342754495, 119409368, -735727453, -968714725 518334522
#> 7    P 0, 0, 0, 342754495, 119409368, -735727453, -968714725 518334522
#> 8    P 0, 0, 0, 342754495, 119409368, -735727453, -968714725 518334522
#>         ...4                   ...5       ...6  ...7 ...8    ...9
#> 1 G, R, n, C ~, g, 6, ',  , *, m, d -451361848 FALSE    0 5249800
#> 2 G, R, n, C ~, g, 6, ',  , *, m, d -451361848 FALSE    0 5249800
#> 3 G, R, n, C ~, g, 6, ',  , *, m, d -451361848 FALSE    0 5249800
#> 4 G, R, n, C ~, g, 6, ',  , *, m, d -451361848 FALSE    0 5249800
#> 5 G, R, n, C ~, g, 6, ',  , *, m, d -451361848 FALSE    0 5249800
#> 6 G, R, n, C ~, g, 6, ',  , *, m, d -451361848 FALSE    0 5249800
#> 7 G, R, n, C ~, g, 6, ',  , *, m, d -451361848 FALSE    0 5249800
#> 8 G, R, n, C ~, g, 6, ',  , *, m, d -451361848 FALSE    0 5249800
any_data_frame(cols = 5L, frac_na = 0.5) |> show_example()
#>   ...1 ...2 ...3       ...4 ...5
#> 1   NA    D   NA 1955-11-28   NA
#> 2   NA    D   NA 1955-11-28   NA
#> 3   NA    D   NA 1955-11-28   NA
#> 4   NA    D   NA 1955-11-28   NA