Random data frame classed object generator

Generate random data frame objects.

Usage

any_data_frame_object(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A quickcheck_generator generator

Examples

any_data_frame_object() |> show_example()
#>      ...1      ...2                ...3       ...4  ...5       ...6  ...7  ...8
#> 1 fQ?-E6'      .HwJ 2462-12-07 10:36:50  886869600  3381 2598-11-30 FALSE  4486
#> 2      8j        +D 2695-05-16 18:14:56   96362232 -2985 2772-06-02 FALSE  2275
#> 3   3x3gn       "W/  714-01-13 01:34:57 -512144611 -2740 2058-05-19 FALSE  1623
#> 4     3<W     <Z5j@ 1875-05-27 23:58:19 -895238069  6834 2030-11-19  TRUE  8020
#> 5      [t <c!GIcS8! 1178-01-17 00:58:48  194975910 -7582 2955-12-24  TRUE -3498
#> 6     Ecg <c!GIcS8! 1938-03-05 11:20:25 -815491690 -3399 1228-01-26  TRUE  1585
#> 7      S}    c%Xi'{ 2367-10-31 10:12:07  478510455     0 1525-10-11 FALSE  1410
#>                  ...9
#> 1  711-10-15 03:54:19
#> 2  844-07-09 11:00:45
#> 3 2835-09-17 23:25:38
#> 4  422-12-15 17:18:57
#> 5 2095-11-10 19:30:12
#> 6  961-05-03 11:07:32
#> 7  203-02-19 17:45:40
any_data_frame_object(rows = 10L) |> show_example()
#>                    ...1
#>  1:  709-05-30 10:46:32
#>  2:   30-03-21 07:21:45
#>  3: 2767-08-07 23:59:03
#>  4:  554-06-24 15:25:02
#>  5: 1237-07-31 09:22:49
#>  6: 1066-06-27 09:16:24
#>  7:  417-02-15 20:29:23
#>  8: 2709-08-08 07:20:53
#>  9: 2093-12-25 13:09:12
#> 10: 2882-05-27 03:46:34
any_data_frame_object(cols = 5L, frac_na = 0.5) |> show_example()
#>                  ...1 ...2       ...3 ...4            ...5
#> 1                <NA>   NA  834216078 TRUE              NA
#> 2 2036-10-04 10:34:48    0 -285755538   NA              NA
#> 3                <NA>   NA         NA TRUE 05:05:55.023320
#> 4                <NA>   NA -624681646 TRUE 14:52:53.857630
#> 5                <NA> 7215  109758209   NA              NA
#> 6                <NA>   NA         NA TRUE              NA
#> 7  452-11-17 11:08:49 3754         NA   NA 18:55:32.385195