Any list generator

Generate lists containing lists or atomic vectors.

Usage

any_list(len = c(1L, 10L), any_na = FALSE)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10. To produce empty vectors set len = 0L or a range like len = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator object.

Examples

any_list() %>% show_example()
#> [[1]]
#> 17:35:23.461366
#> 
#> [[2]]
#> 01:27:51.301445
#> 
#> [[3]]
#> 09:44:04.437234
#> 
any_list(len = 10L, any_na = TRUE) %>% show_example()
#> [[1]]
#> [[1]]$a
#> [1] "2531-01-31" "1970-10-30" "1913-01-24" "1467-10-01" "1245-07-13"
#> [6] "1501-02-03" "1238-01-11" "1147-09-23" "2006-12-07"
#> 
#> [[1]]$b
#> [1] )~   `g7y
#> Levels: )~ `g7y
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] "&3XgBd>,)" "A&"        NA          "9{#"       "/A~C\\)"   "Ez"       
#> [7] "{Z@#W!q@w" "3D0G5"    
#> 
#> [[2]]$b
#> [1] "ZD}zRy'[" "bgqjQHj"  "JMpa0A93"
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1] 5516 6757
#> 
#> [[3]]$b
#> [1] TRUE
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] NA         "K8$nW"    "1OX&MX&<" NA         ".m5~KQ+c" "PWC4tWi" 
#> 
#> [[4]]$b
#> [1] 174613783
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] TRUE TRUE TRUE TRUE   NA
#> 
#> [[5]]$b
#>  [1]    0 7123 2172   NA 7845 1050   NA -249   NA   NA
#> 
#> 
#> [[6]]
#> [[6]]$a
#> 04:54:33.712532
#> 21:22:26.913898
#> 21:44:09.738403
#>              NA
#> 
#> [[6]]$b
#> [1]  430448300  -50937581         NA  805264687         NA  627124660  385726724
#> [8] -424838852  422995117
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1]    NA    NA FALSE    NA    NA
#> 
#> [[7]]$b
#> [1]    NA -5634    NA -7947 -9991 -4173  8107 -5203
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1] 3e -;Ag 0(5lx   y^hJ    1%&     4        s"WSka `dlFg:  =jw4h  
#> Levels: 3e -;Ag 0(5lx y^hJ 1%& 4  s"WSka `dlFg: =jw4h
#> 
#> [[8]]$b
#> 06:47:32.325832
#> 15:00:09.396093
#>              NA
#> 11:33:09.971314
#> 03:52:37.780705
#>              NA
#>              NA
#>              NA
#> 
#> 
#> [[9]]
#> [[9]]$a
#> [1] FALSE    NA  TRUE  TRUE  TRUE  TRUE  TRUE
#> 
#> [[9]]$b
#> [1] "1083-03-20" "2989-05-07" NA           "2284-06-26"
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1]  1 NA
#> 
#> [[10]]$b
#> [1] NA         "UC:'D:"   "Q!BS!w}1" "oc g\""   "E"       
#> 
#>