Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_tibble() |> show_example()
#> # A tibble: 2 x 10
#>   ...1   ...2   ...3     ...4     ...5  ...6  ...7  ...8  ...9            ...10 
#>   <list> <list> <list>   <list>   <lgl> <lgl> <lis> <lis> <time>          <fct> 
#> 1 <dbl … <chr … <named … <named … TRUE  FALSE <dat… <dtt… 09:37:39.789075 "+68j"
#> 2 <fct … <lgl … <named … <named … FALSE TRUE  <dtt… <tim… 07:12:54.181536 "LKTr…
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 x 3
#>           ...1 ...2         ...3      
#>          <dbl> <chr>        <list>    
#>  1   42341963. "Cu "        <dbl [1]> 
#>  2 -645011849. "R8{"        <date [1]>
#>  3  853501696. "03/"        <int [1]> 
#>  4 -768274181. "7YIET'3F\\" <int [1]> 
#>  5 -226019935. "%tQ"        <chr [1]> 
#>  6          0  "luv'{hNN;"  <date [1]>
#>  7 -224427200. "^Bg"        <date [1]>
#>  8  698099674. "|zaUT/S0("  <int [1]> 
#>  9 -932422354. "#wbM"       <time [1]>
#> 10  836903590. "6<,f"       <lgl [1]> 
any_tibble(cols = 5L, frac_na = 0.5) |> show_example()
#> # A tibble: 2 x 5
#>   ...1             ...2      ...3   ...4          ...5
#>   <list>           <chr>     <lgl> <dbl>         <dbl>
#> 1 <named list [2]> NA        FALSE    NA           NA 
#> 2 <named list [2]> SF^+v7U4q TRUE     NA -50369195810.