Any vector generator

Generate atomic vectors or lists.

Usage

any_vector(len = c(1L, 10L), any_na = FALSE)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10. To produce empty vectors set len = 0L or a range like len = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator object.

Examples

any_vector() |> show_example()
#> [[1]]
#> [[1]]$a
#> [1] kaq(t     I-_[/z)r  /:<}60rw4 StvA7=hy2 KR        i!x3{f|f  \\6"el&M 
#> [8] xct0S7    m*;&+"   
#> 9 Levels: kaq(t I-_[/z)r /:<}60rw4 StvA7=hy2 KR i!x3{f|f \\6"el&M ... m*;&+"
#> 
#> [[1]]$b
#> [1]  2545  5151 -5538  5915 -7415  3568
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE  TRUE  TRUE
#> 
#> [[2]]$b
#> [1] "1092-07-14" "2807-11-11" "1819-11-19" "1701-11-06" "1393-08-03"
#> [6] "2187-06-07" "2185-02-02"
#> 
#> 
#> [[3]]
#> [[3]]$a
#> 17:34:01.389557
#> 15:13:17.514320
#> 12:32:12.745895
#> 08:16:35.408213
#> 
#> [[3]]$b
#> [1] "?`ng<"    "wFJ~gPkA"
#> 
#> 
#> [[4]]
#> [[4]]$a
#> 11:33:04.133594
#> 16:40:31.269672
#> 01:43:25.278186
#> 16:30:25.304555
#> 
#> [[4]]$b
#> [1] "':cziH"
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] "2491-01-12" "2871-07-08" "1905-11-11" "2519-11-15" "1672-11-08"
#> [6] "1006-12-12"
#> 
#> [[5]]$b
#>  [1] "1958-04-20" "2977-11-09" "1285-12-21" "1778-06-16" "1269-04-25"
#>  [6] "2314-01-27" "1596-05-20" "1318-10-14" "1513-02-10" "2410-03-31"
#> 
#> 
#> [[6]]
#> [[6]]$a
#> 18:43:51.879842
#> 16:17:30.601449
#> 01:47:40.116220
#> 04:28:22.639137
#> 13:04:40.806364
#> 21:14:38.471750
#> 13:16:50.078618
#> 23:49:23.915865
#> 
#> [[6]]$b
#> [1] "1691-05-22" "1175-12-02" "2565-06-17" "2477-12-28" "2248-10-28"
#> [6] "2685-10-06" "1482-12-10" "1220-03-06"
#> 
#> 
any_vector(len = 10L, any_na = TRUE) |> show_example()
#> [[1]]
#> [[1]]$a
#> [1]  1  1 NA NA  1  1  1  1  1
#> 
#> [[1]]$b
#>  [1]          0  795944285  744855782         NA  998289858 -914205161
#>  [7] -936002489 -797950496   92622818         NA
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] FALSE    NA    NA    NA FALSE FALSE
#> 
#> [[2]]$b
#> [1] : e1! U    
#> Levels: : e1! U
#> 
#> 
#> [[3]]
#> [[3]]$a
#> 19:03:45.821466
#>              NA
#> 01:12:39.111781
#> 19:11:20.648054
#> 17:19:58.873301
#> 
#> [[3]]$b
#> [1] FALSE  TRUE
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] -276582243
#> 
#> [[4]]$b
#> [1] "120-05-21 00:03:27 UTC"  "708-09-18 13:29:52 UTC" 
#> [3] "1304-05-09 20:37:17 UTC" "1928-03-11 18:29:26 UTC"
#> [5] "122-07-29 05:39:11 UTC"  "1652-04-14 19:36:06 UTC"
#> [7] "31-09-25 23:45:03 UTC"  
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1]   NA  169 -778
#> 
#> [[5]]$b
#> 21:40:38.674419
#> 09:44:37.351943
#> 12:49:52.346471
#> 22:10:35.450757
#> 16:15:37.047475
#> 23:49:46.083201
#> 16:40:02.859533
#> 05:19:26.944864
#> 
#> 
#> [[6]]
#> [[6]]$a
#> 10:49:29.742126
#> 06:38:19.206266
#> 15:11:24.864756
#>              NA
#> 
#> [[6]]$b
#> [1]  1  1 NA NA  1  1 NA
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1] "1569-03-10" "2957-03-11" NA           "2554-12-10" "2324-05-13"
#> [6] "1916-07-03" "2063-02-01" "1244-02-09" NA          
#> 
#> [[7]]$b
#> [1] "2091-07-06 06:19:11 UTC" "207-04-12 21:21:09 UTC" 
#> [3] "453-07-04 00:04:12 UTC"  "620-10-19 14:22:08 UTC" 
#> [5] NA                        "1906-10-12 02:35:51 UTC"
#> [7] "1956-07-23 16:39:30 UTC"
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1]  1  1  1 NA  1  1  1  1
#> 
#> [[8]]$b
#>              NA
#>              NA
#> 09:00:07.580548
#> 11:36:27.237525
#>              NA
#> 02:48:22.506770
#> 19:16:56.034206
#> 15:22:50.225023
#> 
#> 
#> [[9]]
#> [[9]]$a
#> 12:08:30.377335
#> 
#> [[9]]$b
#> [1] 410992851 -67696101 -41014827 310189201        NA
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1] "D>"        ".\\%fND"   "K"         "6RAAH|ic`" "`_T:Hwq"   "] P"      
#> [7] NA          ", 95d7=Za" NA         
#> 
#> [[10]]$b
#> 22:56:58.754139
#> 12:17:18.238630
#>              NA
#> 
#>