Any R object generator

Generate any R object. This doesn't actually generate any possible object, just the most common ones, namely atomic vectors, lists, data.frames, tibbles, data.tables, and undefined values like NA, NULL, Inf, and NaN.

Usage

anything(any_empty = TRUE, any_undefined = TRUE)

Arguments

any_empty

Whether empty vectors or data frames should be allowed.

any_undefined

Whether undefined values should be allowed.

Value

A quickcheck_generator object.

Examples

anything() %>% show_example()
#> [[1]]
#> [[1]]$a
#> [1] FALSE  TRUE  TRUE FALSE
#> 
#> [[1]]$b
#> [1] 3]Z|e&A%
#> Levels: 3]Z|e&A%
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] FALSE    NA  TRUE  TRUE    NA FALSE
#> 
#> [[2]]$b
#> 14:03:26.915527
#> 02:20:31.415401
#> 18:22:49.521860
#> 03:32:27.626985
#>              NA
#> 18:25:04.113879
#> 18:49:07.032602
#>              NA
#> 16:51:51.431476
#> 09:30:35.523666
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1] $(Y&_WoV  B@|$J8}   .-ye.jn,  S^{]IgJ-4 M|+-&%N8  '/t)}U=Np
#> Levels: $(Y&_WoV B@|$J8} .-ye.jn, S^{]IgJ-4 M|+-&%N8 '/t)}U=Np
#> 
#> [[3]]$b
#> [1] NA                        "571-11-17 16:10:00 UTC" 
#> [3] "2942-03-22 16:03:06 UTC" "2356-12-05 03:31:17 UTC"
#> [5] "766-04-30 09:42:15 UTC"  NA                       
#> [7] "2645-07-27 19:42:54 UTC" "703-08-28 20:22:45 UTC" 
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] "2025-01-22 14:33:11 UTC" "2766-05-27 22:49:53 UTC"
#> [3] "450-10-13 20:20:31 UTC"  "2015-08-27 19:30:16 UTC"
#> [5] "252-04-18 05:24:49 UTC"  "92-04-10 05:14:55 UTC"  
#> 
#> [[4]]$b
#> [1] -8638 -1469
#> 
#> 
#> [[5]]
#> [[5]]$a
#>              NA
#> 06:31:37.712105
#>              NA
#> 13:54:28.270167
#> 06:50:14.939081
#> 04:58:31.331204
#> 
#> [[5]]$b
#> [1] "496-01-31 12:40:06 UTC"  "168-05-19 06:43:08 UTC" 
#> [3] "1112-11-26 00:03:11 UTC" "1666-12-27 11:06:09 UTC"
#> [5] "69-12-02 18:37:23 UTC"  
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1] "1749-01-17 21:37:07 UTC" NA                       
#> [3] "2330-11-24 04:21:21 UTC" NA                       
#> [5] "2443-10-15 06:56:41 UTC" "2361-10-20 13:33:11 UTC"
#> [7] "2394-10-02 18:55:30 UTC"
#> 
#> [[6]]$b
#> [1]    NA  TRUE  TRUE FALSE FALSE    NA    NA  TRUE
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1] NA
#> 
#> [[7]]$b
#> [1]  TRUE  TRUE FALSE FALSE  TRUE  TRUE
#> 
#> 
#> [[8]]
#> [[8]]$a
#> 09:53:39.338526
#> 
#> [[8]]$b
#> [1] "1166-09-30 01:58:42 UTC" "2495-03-16 06:57:21 UTC"
#> [3] "1282-12-16 13:37:48 UTC" NA                       
#> [5] NA                        "362-04-22 15:06:12 UTC" 
#> [7] NA                        "2778-07-08 10:07:11 UTC"
#> [9] "1654-03-17 09:09:33 UTC"
#> 
#> 
#> [[9]]
#> [[9]]$a
#> [1] "1035-05-12" "2447-02-09"
#> 
#> [[9]]$b
#>  [1] "1211-10-13" NA           NA           "2069-01-31" "1244-10-31"
#>  [6] "1510-05-10" NA           "1495-10-06" "1504-01-29" "1263-04-17"
#> 
#> 
#> [[10]]
#> [[10]]$a
#> NA
#> NA
#> 
#> [[10]]$b
#> [1] NA NA
#> 
#>