data.table generator with randomized columns

data.table generator with randomized columns

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data.table_of(logical_(), date_()) |> show_example()
#>     ...1       ...2  ...3       ...4  ...5       ...6  ...7       ...8
#> 1: FALSE 1026-09-11  TRUE 1346-10-12 FALSE 2732-07-13 FALSE 2580-08-31
#> 2:  TRUE 2675-02-20  TRUE 2599-09-19 FALSE 1879-12-27  TRUE 1522-04-25
#> 3: FALSE 2043-03-05  TRUE 1875-02-22 FALSE 2793-06-14 FALSE 2996-10-02
#> 4:  TRUE 2116-04-21 FALSE 2321-04-22 FALSE 1868-04-10  TRUE 2617-05-07
data.table_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>      ...1       ...2                ...3            ...4                ...5
#>  1:  TRUE  102580403  850-05-18 12:37:49 10:58:36.570288 1916-09-20 05:52:29
#>  2: FALSE          0 1043-10-06 14:41:55 02:39:41.982023 2973-04-19 08:04:16
#>  3:  TRUE          0  904-03-27 22:53:03 18:53:34.119473  494-03-11 19:50:25
#>  4: FALSE  607603505 2246-08-19 12:01:56 15:26:20.035478 2648-03-13 09:56:12
#>  5: FALSE -973416724 1891-07-19 15:11:23 04:09:45.691215 1719-08-24 13:41:54
#>  6:  TRUE -981620195  952-04-18 22:29:17 22:22:56.904518 2811-08-09 13:28:40
#>  7: FALSE  620286971 2522-08-17 14:44:01 07:43:27.313904  626-02-27 10:37:32
#>  8: FALSE -508778438  432-04-03 03:01:10 20:45:49.877657 1635-11-05 04:52:35
#>  9:  TRUE   17729413 1845-03-18 09:21:27 16:29:18.484387 2062-07-22 21:47:55
#> 10: FALSE  300655733 2429-03-07 22:29:32 13:21:14.174491 2726-11-12 00:25:32