Random data frame generator

Random data frame generator

Usage

data_frame_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data frames will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

data_frame_of(logical_(), date_()) |> show_example()
#>    ...1  ...2       ...3       ...4
#> 1 FALSE  TRUE 2835-09-13 1493-09-04
#> 2 FALSE  TRUE 2790-04-29 1822-03-20
#> 3  TRUE FALSE 2954-10-10 2059-07-07
#> 4 FALSE FALSE 1894-03-03 2791-09-21
#> 5  TRUE FALSE 1642-06-24 2971-07-26
#> 6  TRUE  TRUE 1007-05-23 2280-10-03
#> 7  TRUE  TRUE 1737-08-17 2053-11-29
#> 8 FALSE FALSE 2928-04-06 2102-07-03
#> 9 FALSE FALSE 2978-02-15 1233-01-15
data_frame_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>         ...1       ...2  ...3                ...4  ...5
#> 1       pyX7 -878824948 -5791  108-06-11 06:20:35 FALSE
#> 2    &Q+FDrL  302457417   415 2741-08-16 06:53:59  TRUE
#> 3         {(  -10269924 -2586  923-06-27 19:31:26  TRUE
#> 4       lb,p -825371704 -9918 2027-03-06 05:25:06 FALSE
#> 5   'b8AEm(Q -277439768 -7622  159-09-22 11:10:21  TRUE
#> 6       Ef"3          0 -1598  113-01-13 02:03:13  TRUE
#> 7       ynpy          0  5356 1004-02-07 19:59:16 FALSE
#> 8      [B68T -529064248 -3856  681-07-10 20:59:03 FALSE
#> 9   o|mWAStb -945006428 -3868 2049-01-08 20:00:32 FALSE
#> 10 !4z\\dZ5/  716470504   450 2951-09-09 21:53:58  TRUE