Variable length list generator

Generate lists with all values coming from a single generator.

Usage

list_of(generator, len = c(1L, 10L))

Arguments

generator

A quickcheck_generator object.

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10. To produce empty vectors set len = 0L or a range like len = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

list_of(integer_(), len = 10L) %>% show_example()
#> [[1]]
#> [1] -4768   156  8688  -800  8612  5220 -9763  3488  8033
#> 
#> [[2]]
#> [1] -8487 -9857  6688  -313 -6731  6023  2690  -979
#> 
#> [[3]]
#> [1]  4891  9398 -3046  1518  8380 -6081  3516     0
#> 
#> [[4]]
#> [1]  7391 -6129
#> 
#> [[5]]
#> [1] -2334 -6385     0  6758  1742     0
#> 
#> [[6]]
#>  [1]  5213  6609  6262  3596  6313  6675 -3627 -3618 -2143 -3720
#> 
#> [[7]]
#> [1] -8500
#> 
#> [[8]]
#> [1]  2317     0  9138     0 -8993 -8228 -4549  -843
#> 
#> [[9]]
#> [1] 1732
#> 
#> [[10]]
#> [1]  6746 -7462  7757 -1420  6514 -6814
#>