Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 5 × 7
#>   ...1  ...2  ...3       ...4       ...5       ...6  ...7      
#>   <lgl> <lgl> <date>     <date>     <date>     <lgl> <date>    
#> 1 FALSE FALSE 2618-08-07 2348-07-13 1911-03-17 TRUE  1482-07-26
#> 2 TRUE  FALSE 2953-03-12 2648-08-16 2001-01-01 FALSE 2960-02-18
#> 3 FALSE FALSE 2097-03-20 2427-07-24 2516-09-28 FALSE 1178-05-13
#> 4 TRUE  TRUE  2771-02-01 2255-01-03 2133-08-10 FALSE 2240-07-31
#> 5 TRUE  TRUE  2240-11-07 1542-01-27 1625-08-05 TRUE  1589-10-14
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>    ...1                ...2              ...3 ...4              ...5
#>    <dttm>              <date>           <dbl> <date>           <dbl>
#>  1 903-04-29 11:24:13  1009-04-14          0  2870-06-08 -360414416.
#>  2 2888-07-21 06:06:42 2343-08-29  618780365. 2700-10-24    5151833.
#>  3 75-11-13 22:31:42   1020-05-28 -957997578. 1623-05-18 -901799280.
#>  4 1252-06-22 22:15:51 2120-01-31 -852201249. 1999-10-24          0 
#>  5 1669-09-04 09:46:42 2794-09-21 -999843942. 1335-05-17  337325064.
#>  6 432-04-08 13:55:34  1416-06-13  311476966. 2358-03-05  647588427.
#>  7 751-03-11 22:07:29  1960-09-03 -216410729. 2892-12-30 -556772198.
#>  8 2551-12-23 01:24:06 1951-02-23          0  1147-03-20          0 
#>  9 795-07-09 19:11:07  2041-06-07  498426781. 1585-02-21  302159975.
#> 10 784-03-09 06:18:52  1781-02-24  -44647934. 1688-11-14 -507301049.