Random data.table generator

Generate random data.tables.

Usage

any_data.table(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A quickcheck_generator generator

Examples

any_data.table() |> show_example()
#>          ...1       ...2  ...3      ...4      ...5      ...6
#> 1: -178881634  372394059 -4404 <list[2]> L_IstOS'N    /Eeo2:
#> 2: -740203979  671036166  -133 <list[2]>      ,PO'    }*Q0nA
#> 3:  469684683          0  1478 <list[2]> Xrj}?|=DG  Js@^:Xb#
#> 4: -670139383  595903514 -1744 <list[2]>  0EO~#uw` g\\7!dP9%
#> 5:  856727381 -369697797  2793 <list[2]>       PL&        1E
#> 6: -885371196  299052978  5083 <list[2]>      W;\\  YId;&&/Q
#> 7:  100259243  308695939 -9113 <list[2]>  qGaCgAtz      o:o>
any_data.table(rows = 10L) |> show_example()
#>      ...1                ...2      ...3       ...4
#>  1:     0                 991 )tJu9|k:F -969056437
#>  2: -6247                   &    ;61M%N -224443062
#>  3:  7850     23:19:43.899592  DPEBzE/j  803185120
#>  4: -3285          -526410028       /Un          0
#>  5:  1177 2470-12-22 06:04:33        )t  830495779
#>  6: -9935     10:29:51.767621    U[V|uf  418629908
#>  7: -7893          1986-09-01      4nu|  771820170
#>  8:     0           :Oz35FP6e 83Ug{.Uc(  268060312
#>  9:  9971     06:20:52.074096   5o+\\3Q -184408792
#> 10:   622     23:58:20.347302  ZLuCEno$          0
#>                                                                                                                       ...5
#>  1:                                                  1213-12-19,2956-06-24,2368-02-12,2052-02-02,1648-06-11,1531-04-19,...
#>  2:                                                                                                 3342, 2436,-7097,-6808
#>  3:                                                                                FALSE,FALSE, TRUE,FALSE,FALSE,FALSE,...
#>  4:                                                                                                M>=rTw&",WM,>%{'@td,$@T
#>  5:                                                                                                        23:30:57.312205
#>  6:                    20:48:54.560698,13:42:02.627480,02:07:43.921486,07:06:46.953158,06:45:49.195157,20:54:50.199774,...
#>  7:                                        09:38:47.449261,12:04:18.955799,17:24:53.469239,14:09:44.309675,01:20:17.512639
#>  8:                                                           0,-178243719,-431065910,-528368758, 953862865, -45982745,...
#>  9:                                                                                 )?v6,'{rZ[,9,6C!l,sQC\\10&,v-/^r|<,...
#> 10: 2551-05-06 14:32:20,601-05-12 18:03:13,2399-02-25 21:22:52,1165-11-07 09:47:17,2215-10-29 08:10:37,1184-08-09 10:39:24
#>      ...6
#>  1: -3889
#>  2:  5438
#>  3:  9902
#>  4:  -208
#>  5:  7281
#>  6:  9220
#>  7:     0
#>  8:  3634
#>  9:     0
#> 10:  7045
#>                                                                                                    ...7
#>  1:                               -526905155,-663334127, 893026224,-647841808, 339672280,-959737840,...
#>  2:                                                              TRUE,FALSE,FALSE, TRUE, TRUE, TRUE,...
#>  3: 19:02:36.437995,13:24:13.446600,11:43:56.814131,23:34:34.465605,17:10:04.154764,03:34:16.953672,...
#>  4:                     2913-12-15 05:44:47,2929-12-19 19:41:37,2020-08-05 13:43:26,2377-02-06 13:18:05
#>  5:                                                                MY-$6,32eU^vP,x,8eo<,a2V"W;,^J<F,...
#>  6: 19:17:40.974138,22:56:16.862596,04:08:09.364850,21:23:31.844096,11:10:36.574867,05:38:46.021138,...
#>  7:                                        0,-511274650,-738888349,         0, 997415517,-637637538,...
#>  8:                                                                     18:40:16.503721,08:29:20.801741
#>  9:                                                     01:45:28.686991,17:48:03.953328,13:17:54.271814
#> 10:                                                                    1473-09-19,2455-05-05,2243-09-09
#>                ...8       ...9               ...10
#>  1: 20:53:06.917322 -903531355 1308-06-01 08:54:30
#>  2: 17:20:00.988071  605655692 1550-08-31 22:21:52
#>  3: 16:36:59.903862          0 1900-08-01 11:37:29
#>  4: 10:54:03.106258  347384786 1078-08-28 19:34:43
#>  5: 11:38:01.369117  956387919 1111-11-01 13:21:47
#>  6:  12:52:24.39611  546551644   11-04-16 19:50:14
#>  7:  23:58:05.92749 -656836872 1709-03-28 12:27:26
#>  8: 15:29:32.839543 -849657869 2939-03-22 20:55:20
#>  9: 14:32:56.940418 -657239972 1293-07-28 14:57:22
#> 10: 20:23:30.628606  687273473 1881-10-19 02:22:13
any_data.table(cols = 5L, frac_na = 0.5) |> show_example()
#>          ...1                ...2     ...3      ...4      ...5
#> 1: 2661-10-10                  NA     <NA> <list[2]> <list[2]>
#> 2:       <NA> 2156-01-14 07:47:06 v$f=L:cM <list[2]> <list[2]>
#> 3: 2493-12-19 2629-03-28 18:48:21    {A\\_ <list[2]> <list[2]>
#> 4: 1845-12-08                  NA     <NA> <list[2]> <list[2]>
#> 5:       <NA>                  NA cJ[TX~;@ <list[2]> <list[2]>