Random data.table generator

Generate random data.tables.

Usage

any_data.table(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_data.table() |> show_example()
#>                  ...1
#> 1: 844-04-22 20:55:00
#> 2:          663413764
#> 3:         p&\\dnu,_|
#> 4:           G^&1HdD;
any_data.table(rows = 10L) |> show_example()
#>                    ...1  ...2      ...3
#>  1:                   0 -3958      :rH!
#>  2:                sV)U -6582     []~f 
#>  3:            78872905  4099       a0&
#>  4:          1578-11-20 -9410     i<9H@
#>  5: 2116-05-23 07:39:09    94  Bo;QHeXX
#>  6: 2926-07-18 11:18:24   273 l8hk]4=-1
#>  7:  631-02-06 06:46:47     0         )
#>  8: 2537-07-27 13:25:22  6834   1M}^Q%v
#>  9:               -4212  3542     sDe_|
#> 10:            P6l(DQgx -7267       u&Y
#>                                                                             ...4
#>  1:                                                              11:54:23.750185
#>  2: 99-08-08 02:38:16,1217-04-02 13:06:51,810-08-21 20:22:15,2679-01-12 23:40:41
#>  3:        1144-03-29,2766-11-05,1987-11-01,2834-12-19,1731-01-18,1665-04-15,...
#>  4:                                             2662-05-09,2681-04-11,1672-02-21
#>  5:                                              06:16:59.467046,04:59:39.600130
#>  6:                                                      sNTwGQn,,^:+,+4,o6T@!=N
#>  7:                                                                U1T,Psk,\\BWG
#>  8:        2674-11-22,2287-08-09,1125-03-29,2434-09-18,1213-01-06,1622-10-15,...
#>  9:                    446-08-25 06:06:06,1418-04-20 06:10:50,875-01-19 11:41:38
#> 10:                                       657-08-22 05:46:40,2077-03-27 18:20:39
#>                    ...5  ...6       ...7       ...8
#>  1:                   0 -8246        PTN 2898-01-09
#>  2:      09:21:12.60835  9861 Y\\$YmLW=M 1206-01-31
#>  3:               -8607  7429  Gdp6MnR 3 2755-04-17
#>  4:          1950-03-08     0         '. 1646-05-19
#>  5:                TRUE -4355         BF 1098-11-27
#>  6: 1713-08-30 17:41:42 -9474    Z'K'@tj 1464-09-08
#>  7:            -8634596  -425       rmEd 2410-04-24
#>  8:                TRUE  3624          # 2773-02-18
#>  9:               FALSE -9001         ?> 1489-02-21
#> 10:                TRUE -9573    9c6a\\= 2282-01-30
any_data.table(cols = 5L, any_na = TRUE) |> show_example()
#>    ...1 ...2 ...3            ...4                         ...5
#> 1: <NA> TRUE   NA 07:04:21.840074 NA,aJY,3Uw,8,NA,t*8x=K=0,...