Random data.table generator

Generate random data.tables.

Usage

any_data.table(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_data.table() |> show_example()
#>                   ...1       ...2
#> 1: 2785-02-23 18:40:28 1352-08-20
#> 2: 2147-07-18 00:05:41 2248-02-03
#> 3:  662-12-16 18:16:00 2347-11-26
#> 4: 1039-09-02 14:13:19 1407-05-14
#> 5:  440-05-06 06:50:59 1204-02-07
#> 6: 2138-02-25 20:15:32 1974-10-12
#>                                                                                                                          ...3
#> 1:                                                                                 y-(72Ai/,W-f,q*U,?=hrqKCq`,_M kLd,W(8),...
#> 2:                                                                                     st@ZF,moz,'VM)Y,"~|I>U6,),3\\EH1F$,...
#> 3:      625-08-07 07:23:48,1956-04-18 13:22:55,550-09-27 07:45:46,2720-06-21 05:51:57,2796-10-30 23:59:33,1916-11-16 07:34:35
#> 4:                                                                                           1680-09-03,1615-05-29,2324-03-09
#> 5:                                                                                                       195823098,-294752867
#> 6: 951-11-20 00:55:41,2932-04-23 09:47:08,1245-12-25 10:53:19,2188-02-15 04:41:13,1136-09-17 05:48:19,2821-12-28 06:57:23,...
#>     ...4      ...5      ...6      ...7      ...8
#> 1:     0 <list[2]> <list[2]> <list[2]> <list[2]>
#> 2:  4367 <list[2]> <list[2]> <list[2]> <list[2]>
#> 3:  -523 <list[2]> <list[2]> <list[2]> <list[2]>
#> 4: -2882 <list[2]> <list[2]> <list[2]> <list[2]>
#> 5: -5206 <list[2]> <list[2]> <list[2]> <list[2]>
#> 6: -7297 <list[2]> <list[2]> <list[2]> <list[2]>
any_data.table(rows = 10L) |> show_example()
#>          ...1      ...2            ...3                ...4
#>  1:     hZ!%5    OT=UD/ 07:15:33.609202 2921-10-04 20:11:49
#>  2:      s4Kl  0gN${;8u 14:02:33.102943 1590-11-21 16:26:16
#>  3:         y  F2[TBz1( 18:40:33.142323  361-01-20 00:05:57
#>  4:  qZ5),.@- #.sXi3}0T 05:41:18.149733 2222-02-19 15:27:16
#>  5:        P5  fi<t\\\\ 15:28:43.264305 1131-06-19 16:26:01
#>  6:      ?r!_ fk;m w:yR 04:26:20.233553 2424-06-30 16:37:50
#>  7:     *e9$y        nf 14:44:43.335109 2045-02-02 22:47:54
#>  8:         L         F 05:46:58.141600 2746-05-19 21:39:24
#>  9: (uJ6s<MVG      '^V- 19:05:13.547656 2023-11-16 23:17:21
#> 10:         R    )mwuh: 14:38:48.925865 1758-05-26 23:51:58
any_data.table(cols = 5L, any_na = TRUE) |> show_example()
#>          ...1      ...2            ...3      ...4      ...5
#> 1: -416368529 878497571 04:13:45.798942 <list[2]> <list[2]>
#> 2:         NA        NA 03:02:58.948319 <list[2]> <list[2]>
#> 3:  543733314        NA 23:26:39.701577 <list[2]> <list[2]>
#> 4:         NA      Y7$9              NA <list[2]> <list[2]>
#> 5:          0        NA 23:27:13.785387 <list[2]> <list[2]>