Random data.table generator

Generate random data.tables.

Usage

any_data.table(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_data.table() |> show_example()
#>          ...1      ...2      ...3       ...4  ...5      ...6               ...7
#> 1: 2560-04-25 <list[2]>        NY  482896030     0      Z'W9         2718-08-31
#> 2: 1500-12-25 <list[2]> 3C(xpsp[8 -238337170  5826       f''                 R%
#> 3: 1698-11-27 <list[2]>         I  892651975  4572         b 989-06-09 09:54:12
#> 4: 1773-04-05 <list[2]>  w^^PXZ[h  453564791 -4053 yc07|(4@L              FALSE
#> 5: 1768-12-08 <list[2]>     0epc*  460571848 -4292  j]djK)wa              -8045
#>                   ...8
#> 1:  654-12-18 17:08:16
#> 2: 1710-08-04 00:12:44
#> 3: 2871-05-28 02:27:53
#> 4:  466-05-28 18:42:20
#> 5: 2235-09-06 05:55:24
any_data.table(rows = 10L) |> show_example()
#>                ...1            ...2                ...3            ...4
#>  1: 09:35:25.361628 08:11:08.568587               FALSE 04:17:28.878473
#>  2: 04:52:04.303625 10:10:16.307397     19:11:06.074051 08:06:19.140012
#>  3: 03:04:05.758066 14:12:17.285087                TRUE 02:57:58.159125
#>  4: 16:49:50.180187 14:13:59.192910                %P8R 10:06:19.782088
#>  5: 11:47:35.078798 10:50:58.026041             Qt+OC.I 12:28:48.418248
#>  6: 17:10:21.623990 16:21:09.544760             |HekO-F 23:55:31.598189
#>  7: 21:46:54.854627 19:29:47.637022 1917-02-28 02:39:17 13:57:11.737696
#>  8: 08:14:49.712585 03:49:41.384388 1574-08-02 10:58:59 13:20:19.917216
#>  9: 01:17:02.431907 04:49:23.968833               -9608 09:19:36.037170
#> 10: 21:28:57.698726 20:49:53.520549                TRUE 10:30:14.365266
#>                                                                                                    ...5
#>  1:                               -978780262,  41487678, 350743391,         0,-203630330,-995697550,...
#>  2:                                                              TRUE,FALSE, TRUE, TRUE,FALSE,FALSE,...
#>  3: 07:02:36.138169,16:14:43.069535,04:25:39.055033,08:01:54.902963,17:13:47.768303,12:44:43.227364,...
#>  4:                                                                                               FALSE
#>  5:                                                                                         0, 969,5840
#>  6:                                                              TRUE, TRUE, TRUE,FALSE,FALSE,FALSE,...
#>  7:                                                                                                   k
#>  8:                                -77751604, -11492531,-710909377,-198912622, 871476698,-185618194,...
#>  9:                                                           r&pTz,n,p32o,u,Lq?%,!H4[.=k7=,f2`D0N,,...
#> 10:                                                               YWe<,495A,MB#P.W,Oa]OimF,K9MAU,t ,...
any_data.table(cols = 5L, any_na = TRUE) |> show_example()
#>               ...1            ...2            ...3            ...4      ...5
#> 1:      1105-09-22      -971009246 20:23:31.799336 03:18:24.260170 <list[2]>
#> 2: 23:33:13.524776      1157-05-04 11:30:10.372778 18:20:01.197221 <list[2]>
#> 3:           -2276           -9578 12:21:13.577236 23:58:13.718084 <list[2]>
#> 4:      -585406379       m-6;A(M^) 01:29:19.671075 10:01:06.624928 <list[2]>
#> 5:      1143-09-28           FALSE 20:41:09.850693              NA <list[2]>
#> 6:              $# 18:45:58.318952              NA 09:35:05.969200 <list[2]>