Any data frame generator

Generate data.frames.

Usage

any_data_frame(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator object.

Examples

any_data_frame() |> show_example()
#>   ...1  ...2      ...3    ...4      ...5  ...6                ...7
#> 1 !z\\ FALSE 702051592 ]@2~%*P .U$n;;- < FALSE 2258-04-23 02:55:18
#> 2    ^ FALSE -18914399  6o*2X(      %QDw FALSE 1888-11-14 09:40:14
#> 3    } FALSE         0    7ux[     NgtF<  TRUE 1540-03-24 03:34:33
#>              ...8     ...9    ...10
#> 1 18:22:35.360181        5      D0k
#> 2 08:42:43.267521  zC%8JZ[ lo~b o ;
#> 3 06:14:52.495235        / iSBUd$bM
any_data_frame(rows = 10L) |> show_example()
#>          ...1  ...2  ...3
#> 1   159474846 FALSE -4220
#> 2   550488148  TRUE  9033
#> 3   -86534242  TRUE     0
#> 4   779842298 FALSE -1012
#> 5   884814916  TRUE -4572
#> 6   177343578  TRUE  6192
#> 7   -55076904  TRUE -6137
#> 8  -164357371 FALSE -6468
#> 9  -436940702 FALSE -7849
#> 10  102363877 FALSE -8011
any_data_frame(cols = 5L, any_na = TRUE) |> show_example()
#>         ...1    ...2       ...3 ...4 ...5
#> 1  448823796    Bsjm -975109119   NA   NA
#> 2         NA ePk{v:K  178685572    1 TRUE
#> 3         NA  ?UqaX$  153062338   NA TRUE
#> 4 -891206247    <NA> -994794827    1   NA
#> 5 -755417289    <NA>         NA    1 TRUE
#> 6  967290500    <NA> -129503794   NA TRUE
#> 7 -300919800 7mg+\\a -614184054    1 TRUE
#> 8  352751359 y/K7,!z -651591839    1   NA