Random data frame generator

Generate random data frames.

Usage

any_data_frame(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_data_frame() |> show_example()
#>              ...1       ...2       ...3            ...4       ...5  ...6
#> 1 11:41:28.165763 2801-09-30 1480-10-17 10:28:50.068429 1652-05-15  TRUE
#> 2 07:06:47.637455 2431-06-19 1873-01-06 09:33:00.134787 2819-09-08 FALSE
#> 3 15:37:58.968015 2992-01-18 1501-01-24 08:03:59.663526 1489-07-27 FALSE
#> 4 01:46:42.100751 2027-02-22 2282-02-16 02:14:19.084695 1119-04-16 FALSE
#> 5 15:25:54.409521 1074-10-27 2495-03-07 09:06:28.652163 2148-09-01 FALSE
#> 6 22:39:49.649524 2900-11-11 1296-01-26 11:11:05.422113 2822-05-01 FALSE
#> 7 15:47:13.914144 2687-07-05 2590-03-24 23:33:02.222152 2228-10-07 FALSE
#> 8 16:51:30.081084 1155-09-19 1509-03-20 22:12:24.311932 1311-08-14  TRUE
#> 9 15:32:02.727339 1642-03-08 1212-09-02 05:35:10.124285 1708-10-25 FALSE
#>         ...7      ...8                ...9
#> 1  558648035   7<z}^52 1500-11-08 08:27:27
#> 2  711220366         $  354-08-10 03:02:47
#> 3  697378979      $2nw 2986-10-22 23:03:34
#> 4 -464457138      M;S9  410-04-22 03:25:48
#> 5 -540170097         )  181-06-04 20:40:55
#> 6 -818459970         Y 1548-11-25 06:12:36
#> 7 -604454241    1-'T2G 2199-09-24 20:42:21
#> 8          0 H>"sug_I(  972-11-26 10:02:43
#> 9 -661288347 mm~ro!*T@ 1813-02-24 06:13:56
any_data_frame(rows = 10L) |> show_example()
#>                   ...1  ...2  ...3     ...4
#> 1   497-09-28 02:10:13  TRUE   935  JVR}W I
#> 2  1610-01-10 21:45:18  TRUE -4538 akzU%;&'
#> 3  1015-07-05 04:32:04  TRUE -8589  yIB%NYi
#> 4  2798-01-05 22:14:06 FALSE  2555    SzjM,
#> 5  1467-01-15 19:36:53  TRUE  6574     ,4;L
#> 6   975-05-23 07:28:56 FALSE -8208 >M(8Y?Sc
#> 7  1138-02-23 12:42:43  TRUE  6368        +
#> 8  2875-09-19 06:20:42 FALSE -8137     zbj-
#> 9   801-02-06 13:48:13 FALSE -3472        ~
#> 10 2650-10-31 22:40:09 FALSE     0 m}|/V{DG
any_data_frame(cols = 5L, any_na = TRUE) |> show_example()
#>                  ...1 ...2       ...3 ...4      ...5
#> 1 1641-07-18 01:36:32   NA 1304-05-01   NA 326957435
#> 2 2706-05-01 04:41:40    1 2630-06-05    1  19692099