Random data frame generator

Generate random data frames.

Usage

any_data_frame(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A quickcheck_generator generator

Examples

any_data_frame() |> show_example()
#>              ...1 ...2 ...3       ...4                ...5 ...6  ...7
#> 1 01:26:46.856513   v1    0 2041-06-07 2204-09-10 10:53:43    = FALSE
#>         ...8    ...9
#> 1 -838304528 Be!>9H+
any_data_frame(rows = 10L) |> show_example()
#>                   ...1            ...2                ...3                ...4
#> 1  1616-08-05 01:27:32 13:32:07.746987 1715-03-17 13:08:55   45-09-20 20:39:24
#> 2  2045-12-30 00:43:44 09:55:03.001498 2730-08-26 12:39:23 2222-05-30 15:33:12
#> 3  2017-12-25 02:20:42 00:01:05.499188 1207-06-10 12:10:39 2306-12-30 07:58:20
#> 4  1551-05-08 17:53:49 19:23:25.341690 1508-12-25 07:28:44 1762-08-08 02:35:50
#> 5  2515-07-23 12:28:04 07:52:55.801465 2725-10-03 03:47:15 2498-01-23 20:07:10
#> 6   573-06-08 16:29:45 06:54:49.601561 2444-11-04 22:53:17 2296-10-25 18:37:17
#> 7  2640-03-24 01:28:28 00:11:20.882532  313-03-02 04:45:58 2617-02-27 02:33:11
#> 8  1883-05-06 16:55:16 14:21:24.369793 1261-07-11 09:56:49  167-09-21 10:12:10
#> 9  1518-05-04 21:02:57 00:07:20.114104 2389-04-29 15:06:12  370-04-19 03:18:41
#> 10  626-01-25 05:10:43 16:04:02.668854 2833-12-08 07:16:21  709-09-24 09:21:54
#>     ...5
#> 1  FALSE
#> 2  FALSE
#> 3   TRUE
#> 4   TRUE
#> 5   TRUE
#> 6   TRUE
#> 7  FALSE
#> 8   TRUE
#> 9   TRUE
#> 10  TRUE
any_data_frame(cols = 5L, frac_na = 0.5) |> show_example()
#>              ...1    ...2            ...3  ...4            ...5
#> 1              NA    <NA> 20:40:47.092883    NA              NA
#> 2              NA    OkQj              NA    NA              NA
#> 3              NA      Kt              NA FALSE 08:32:47.762033
#> 4              NA !8#]FWA 07:05:46.739778    NA              NA
#> 5 10:46:02.125156      _9 00:52:07.672966    NA 05:08:04.337296
#> 6 11:39:34.898205    <NA> 09:02:34.453308    NA              NA