Random data frame generator

Generate random data frames.

Usage

any_data_frame(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_data_frame() |> show_example()
#>         ...1            ...2       ...3 ...4       ...5     ...6    ...7
#> 1 1760-02-05 16:07:04.258324 -812630672 5006 -680356258       s6 GdT0WuX
#> 2 1999-07-21 07:25:38.820601  167036919 4722 -848937520 (,.GDb7#  a-l"mM
#>        ...8    ...9
#> 1 *Zp[9j*{, k&v)[-q
#> 2         %      st
any_data_frame(rows = 10L) |> show_example()
#>       ...1      ...2       ...3  ...4                ...5            ...6
#> 1    potw,      >ZPN 1310-06-05  TRUE  603-02-26 10:09:02 15:53:32.621466
#> 2   9[BP13     GLe+r 1236-05-04 FALSE 2841-10-08 08:40:28 13:36:08.174020
#> 3  r&4Cm%x       ()t 2540-01-08  TRUE 1279-04-24 18:36:23 23:47:05.305407
#> 4     hb w {BWjHlA=Y 1445-07-21 FALSE 2593-03-26 16:17:24 08:31:23.394869
#> 5    D?3=T  'TM^Rf>L 1332-01-30 FALSE 1022-10-31 03:10:38 22:18:25.655397
#> 6     6yL&      VSYX 2709-01-07 FALSE  159-09-04 19:00:41 07:04:36.146148
#> 7        D       C!v 2564-09-18  TRUE   45-08-06 22:46:36 11:36:34.668981
#> 8  |xto1ME     xr${` 1533-02-27 FALSE   72-04-28 21:04:10 05:34:16.918712
#> 9        z         K 2168-12-06 FALSE 2992-02-12 06:00:02 03:37:45.480504
#> 10       k t5r{O|\\[ 1704-01-11  TRUE 1643-08-11 10:20:49 08:49:04.508274
#>       ...7       ...8
#> 1    P=?5B 1181-05-09
#> 2  x;7P>.F 2131-10-04
#> 3        R 2301-01-15
#> 4       V& 2106-04-21
#> 5      E V 2635-06-05
#> 6   l1B`7} 1605-05-02
#> 7     @TV! 1817-06-17
#> 8      E V 2681-06-06
#> 9     ""?Q 1572-04-12
#> 10     k%y 2743-08-29
any_data_frame(cols = 5L, any_na = TRUE) |> show_example()
#>              ...1            ...2     ...3 ...4  ...5
#> 1 20:38:42.751786 03:26:19.257749     Rz~_ TRUE -2625
#> 2              NA 15:16:45.351362 uXL|St6"   NA     0