Random data frame generator

Generate random data frames.

Usage

any_data_frame(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_data_frame() |> show_example()
#>                  ...1            ...2                ...3       ...4  ...5
#> 1 1866-09-20 07:01:03 02:27:01.246754 1979-03-12 03:48:30 2732-04-12 -9936
#> 2  383-12-09 00:07:50 17:54:46.973816  384-08-13 23:45:27 2095-11-29  1523
#> 3 2823-12-18 23:23:29 06:19:19.266983 2661-01-08 21:55:43 2335-11-22 -3477
#> 4 1779-05-02 02:32:24 19:59:18.383366  641-06-04 07:15:46 1507-06-25  1474
#>         ...6  ...7                ...8
#> 1 2196-04-27  ~`v} 2570-06-15 07:08:29
#> 2 2223-03-12   Qv' 1731-03-10 02:38:31
#> 3 2195-04-26   _vy 1281-10-25 00:27:12
#> 4 1733-01-21 MCguV  795-07-11 19:14:59
any_data_frame(rows = 10L) |> show_example()
#>          ...1                ...2            ...3       ...4       ...5
#> 1   785784288  931-12-11 03:45:18 05:39:47.343773 -163499505  470458474
#> 2  -122858027 1343-01-22 16:50:14 08:02:07.024262  805108983 -535690533
#> 3   111147834 1522-01-28 04:09:42 03:47:28.892273  123628598   23675284
#> 4   160530772  246-06-30 15:08:34 22:26:07.661505  163805657 -427892938
#> 5   362633883 2672-07-21 05:19:50 23:11:17.471488 -735424105  433403514
#> 6  -906943741 1942-01-23 04:36:24 02:54:01.834516 -464804641  749418189
#> 7  -382971806  521-02-08 16:39:44 21:49:03.672212  893105294  545428692
#> 8           0 1094-10-24 17:58:22 18:01:36.045840  863712176  -42788826
#> 9    11141971  464-08-06 05:03:38 14:31:43.392483 -120676725  -81998154
#> 10 -846281414 1160-12-21 07:49:24 15:48:41.637698 -117642990  584724120
#>         ...6       ...7   ...8
#> 1       ac*9 2490-08-08    !W7
#> 2          6 2009-03-10   Cz]y
#> 3          + 1789-11-20 o)sFN9
#> 4    @Au<]^+ 2040-01-14      /
#> 5       Y|-R 2519-12-18   c#0 
#> 6     '/WMgK 2100-01-07     e#
#> 7      f*wD& 1427-06-15  .)lF;
#> 8         0s 2172-08-14      g
#> 9    ue]+;P% 2979-07-24    ogy
#> 10 [zW@p.3<g 1176-06-30  t)-[7
any_data_frame(cols = 5L, any_na = TRUE) |> show_example()
#>              ...1       ...2  ...3                ...4     ...5
#> 1 11:45:53.187772 -791655505 FALSE 2656-01-13 08:28:29 rcTDwWJS
#> 2 14:34:29.313420 -354491960 FALSE                <NA>      O+5
#> 3 02:51:17.348336         NA FALSE 2299-08-01 20:38:54      /X0