Random list generator

Generate random lists containing other lists or atomic vectors of variable lengths.

Usage

any_list(len = c(1L, 10L), frac_na = 0)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_list() |> show_example()
#> [[1]]
#> [[1]]$a
#> [1] "935-11-11 11:23:05 UTC"  "1223-07-19 06:32:19 UTC"
#> [3] "914-12-02 14:33:07 UTC"  "2133-03-20 18:26:18 UTC"
#> [5] "2778-10-14 01:04:48 UTC" "2668-11-17 18:39:18 UTC"
#> [7] "2617-03-23 16:56:14 UTC" "1572-01-21 03:17:44 UTC"
#> 
#> [[1]]$b
#> [1] ^!RGP.Gn b$.      EF"(mZI.
#> Levels: ^!RGP.Gn b$. EF"(mZI.
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1]  TRUE FALSE  TRUE
#> 
#> [[2]]$b
#> [1]  2679  9888  2841 -4340  1522  7045 -6372
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1]  1779 -7189  2160 -8157  9941 -3274     0
#> 
#> [[3]]$b
#> [1] "2302-09-13"
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] URrS-O   ~0@=     ?h0tK'   ,,en0oc} -qQ\\    U8v&     YB@       -z'    
#> Levels: URrS-O ~0@= ?h0tK' ,,en0oc} -qQ\\ U8v& YB@  -z'
#> 
#> [[4]]$b
#> [1] -1319  8686  4394  1857 -6367     0 -6874
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1]  525013277 -977749616   40664243 -253842213  115380964
#> 
#> [[5]]$b
#> [1] -74611965 126888010 943412577         0 312833773 168576314         0
#> [8]         0
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1] "1209-11-08 11:49:11 UTC" "1144-09-10 15:16:41 UTC"
#> [3] "2894-06-17 15:25:49 UTC"
#> 
#> [[6]]$b
#> [1] -335229307  407530667  459130418  569809161  345967485
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1] -659746375  909051037
#> 
#> [[7]]$b
#> [1] ly>i|?:&' |         uM<5      C\\<,n)   *IP}14.L  },=       >nS7nON6h
#> [8] 7h4,U~@S  'KGu     
#> Levels: ly>i|?:&' | uM<5 C\\<,n) *IP}14.L },= >nS7nON6h 7h4,U~@S 'KGu
#> 
#> 
any_list(len = 10L, frac_na = 0.5) |> show_example()
#> [[1]]
#> [[1]]$a
#> [1]    NA 48208
#> 
#> [[1]]$b
#>  [1]    NA    NA  TRUE  TRUE FALSE FALSE  TRUE    NA  TRUE  TRUE
#> 
#> 
#> [[2]]
#> [[2]]$a
#>              NA
#>              NA
#> 14:48:53.470327
#>              NA
#> 12:00:40.603365
#>              NA
#> 
#> [[2]]$b
#> [1]      NA  -97385      NA  347350      NA -344894      NA
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1]           NA           NA  17479236739 -56283912185
#> 
#> [[3]]$b
#>  [1] " {zB9:O"   NA          "'(DGC)59+" NA          ","         "N<mp:"    
#>  [7] "\\"        NA          "|^HR"      NA         
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] -798076998         NA  623093582
#> 
#> [[4]]$b
#> [1] "2335-06-13" "2932-12-14" NA           NA           "2201-02-27"
#> [6] NA           NA           NA          
#> 
#> 
#> [[5]]
#> [[5]]$a
#>              NA
#> 13:23:30.114345
#>              NA
#>              NA
#> 00:33:58.152319
#> 22:27:13.512427
#> 14:25:07.282680
#>              NA
#>              NA
#> 
#> [[5]]$b
#> [1] "fyH/s\\^T," "4|cSF8"     "0dx"        NA           "C{/+J2>)"  
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1]  1 NA NA  1
#> 
#> [[6]]$b
#> [1] FALSE    NA    NA  TRUE
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1] "1421-07-22" NA           "2327-07-17" NA           NA          
#> [6] "2766-03-23" NA           "2355-03-12" "1575-08-26"
#> 
#> [[7]]$b
#> [1]  593109351         NA -791912520
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1]         NA 9646047667         NA
#> 
#> [[8]]$b
#> [1] NA
#> 
#> 
#> [[9]]
#> [[9]]$a
#> 06:20:12.43862
#> 
#> [[9]]$b
#> [1] "726-08-31 16:21:26 UTC" NA                       NA                      
#> [4] NA                       NA                       NA                      
#> [7] NA                       NA                       NA                      
#> 
#> 
#> [[10]]
#> [[10]]$a
#>              NA
#>              NA
#> 14:44:15.321343
#>              NA
#>              NA
#> 06:31:05.573444
#> 22:42:52.657980
#> 07:15:42.843344
#>              NA
#> 
#> [[10]]$b
#> [1] "'"       NA        "-LtIKqt"
#> 
#>