Any list generator

Generate lists containing lists or atomic vectors.

Usage

any_list(len = c(1L, 10L), any_na = FALSE)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10. To produce empty vectors set len = 0L or a range like len = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator object.

Examples

any_list() %>% show_example()
#> [[1]]
#> [1] 628831436
#> 
#> [[2]]
#> [1] 384775091
#> 
#> [[3]]
#> [1] 0
#> 
#> [[4]]
#> [1] 0
#> 
#> [[5]]
#> [1] 771838428
#> 
#> [[6]]
#> [1] -407021579
#> 
#> [[7]]
#> [1] -813578981
#> 
#> [[8]]
#> [1] -461272501
#> 
#> [[9]]
#> [1] 59924764
#> 
#> [[10]]
#> [1] 579012814
#> 
any_list(len = 10L, any_na = TRUE) %>% show_example()
#> [[1]]
#> [[1]]$a
#> [1] -2896    NA -9157 -2057     0 -9221  9747    NA
#> 
#> [[1]]$b
#> [1] "ANg&r" "se"   
#> 
#> 
#> [[2]]
#> [[2]]$a
#> 17:01:50.652189
#>              NA
#> 15:14:17.022964
#> 02:16:16.171540
#> 
#> [[2]]$b
#> [1] NA NA
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1]  TRUE FALSE    NA  TRUE    NA  TRUE  TRUE FALSE FALSE
#> 
#> [[3]]$b
#>  [1]  TRUE  TRUE  TRUE    NA  TRUE  TRUE FALSE FALSE  TRUE  TRUE
#> 
#> 
#> [[4]]
#> [[4]]$a
#>  [1] "1437-12-10" "1578-01-09" "2467-06-29" NA           "1415-11-03"
#>  [6] "1705-03-10" "2042-11-13" "1908-10-01" "1324-11-07" "1376-11-01"
#> 
#> [[4]]$b
#> [1] NA                        "314-08-16 04:45:16 UTC" 
#> [3] "314-12-16 18:29:14 UTC"  "1877-01-04 04:02:47 UTC"
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] "1048-10-24 05:15:34 UTC" "2693-10-25 11:32:28 UTC"
#> [3] "719-10-25 12:18:11 UTC" 
#> 
#> [[5]]$b
#>  [1]  -566  7393  4055    NA   496 -2242     0 -4262    NA  1322
#> 
#> 
#> [[6]]
#> [[6]]$a
#>  [1] -855368942          0         NA -854601259         NA -663214189
#>  [7]  745814596 -673814842 -472609983 -670267591
#> 
#> [[6]]$b
#> [1] -1844868       NA
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1] FALSE
#> 
#> [[7]]$b
#> [1] -1875
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1] FALSE FALSE  TRUE  TRUE  TRUE    NA    NA  TRUE    NA
#> 
#> [[8]]$b
#>  [1] NA           "1744-09-27" "2697-11-10" "1953-05-17" "2401-06-01"
#>  [6] "1252-11-19" "1307-12-17" NA           NA           "2136-03-17"
#> 
#> 
#> [[9]]
#> [[9]]$a
#> 22:02:30.073572
#> 23:13:05.463798
#>              NA
#> 11:05:06.402105
#> 20:25:28.356438
#> 16:31:07.022051
#> 
#> [[9]]$b
#> 11:20:44.164861
#> 15:46:50.743526
#> 00:59:47.713366
#>              NA
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1] NA                        "2893-09-28 11:59:10 UTC"
#> [3] NA                        "2062-12-29 04:03:57 UTC"
#> [5] NA                        "2516-09-17 11:12:53 UTC"
#> 
#> [[10]]$b
#> [1] NA         "Fd/7I8"   NA         "JtI1"     ":3O"      "/|"       NA        
#> [8] "GV;z<$a9" "F}<i"    
#> 
#>