Random list generator

Generate random lists containing other lists or atomic vectors of variable lengths.

Usage

any_list(len = c(1L, 10L), any_na = FALSE)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10. To produce empty vectors set len = 0L or a range like len = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_list() |> show_example()
#> [[1]]
#> [[1]]$a
#> [1] "Kxm5I/6$|"
#> 
#> [[1]]$b
#> 00:11:13.485011
#> 16:30:22.531880
#> 
#> 
#> [[2]]
#> [[2]]$a
#>  [1] -973452537 -732626508 -776317258 -561020009 -449432665  407730310
#>  [7]  303159812 -418316975 -262554333   27910966
#> 
#> [[2]]$b
#> [1] aWr#gx%P! sJIZ<O    r4DoFt,R  (fJ\\_1ob #T       
#> Levels: aWr#gx%P! sJIZ<O  r4DoFt,R (fJ\\_1ob #T
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1] "1137-09-27 15:04:54 UTC" "622-07-08 18:50:17 UTC" 
#> [3] "1201-10-23 20:22:00 UTC" "780-10-06 02:37:47 UTC" 
#> [5] "492-10-24 17:28:04 UTC"  "946-09-18 00:29:17 UTC" 
#> [7] "2553-09-19 06:30:57 UTC" "2831-07-08 17:59:57 UTC"
#> [9] "2777-12-07 09:39:00 UTC"
#> 
#> [[3]]$b
#> [1] "740-10-18 04:16:26 UTC" "837-09-08 08:00:32 UTC" "499-02-22 19:09:37 UTC"
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1]  4654  7321  5267  7425  1348  6969  -280  9573 -5741
#> 
#> [[4]]$b
#> 10:31:56.331145
#> 19:48:50.221101
#> 00:50:17.571579
#> 07:02:34.399877
#> 22:14:03.243121
#> 00:09:50.501167
#> 10:40:19.450273
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] z}Vc7jlV
#> Levels: z}Vc7jlV
#> 
#> [[5]]$b
#> [1] QORW      5 *Ohs    <.WBP5fc  {0#u$     ?,K^p0zsG l0k      
#> Levels: QORW 5 *Ohs <.WBP5fc {0#u$ ?,K^p0zsG l0k
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1] "'"         "R"         "4H3ds#s]w" " _MKGoZ"  
#> 
#> [[6]]$b
#> [1] -484314871 -558888803 -365625353          0  619235686 -760433067   81462707
#> 
#> 
#> [[7]]
#> [[7]]$a
#> 04:16:25.561324
#> 18:36:54.178473
#> 
#> [[7]]$b
#> [1] "n^KFFOa." "5dT"      "59"       "fCCw 2,Q"
#> 
#> 
#> [[8]]
#> [[8]]$a
#>  [1]  379471943 -728301639  902285529          0  -47264911          0
#>  [7]   13891583  112926130 -179193070          0
#> 
#> [[8]]$b
#> [1] -273607356 -537365389  997512377 -867025330          0
#> 
#> 
any_list(len = 10L, any_na = TRUE) |> show_example()
#> [[1]]
#> [[1]]$a
#> 03:51:30.899686
#>              NA
#> 
#> [[1]]$b
#> [1]  TRUE    NA FALSE FALSE  TRUE    NA    NA
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] NA  1
#> 
#> [[2]]$b
#> [1] NA          "J"         "(mVBX&a''" "D"         NA          "s\"P@kdL" 
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1] FALSE FALSE    NA  TRUE  TRUE    NA  TRUE
#> 
#> [[3]]$b
#> 09:51:48.666137
#> 12:56:49.279175
#> 21:39:48.227137
#> 05:21:29.265436
#> 19:25:42.139787
#> 22:15:30.988521
#> 00:25:25.007914
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] NA                       "586-09-05 06:09:07 UTC"
#> 
#> [[4]]$b
#> [1]  -357    NA  8008  3595  8373 -8765     0    NA
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] "d[Y+I?AR8"  "E+X%\""     "1vMw!(f#"   "H[}L50:\""  "Td>A7N8\\P"
#> [6] "?Vt-]"      "CW\""       "H^r"       
#> 
#> [[5]]$b
#> [1] TRUE
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1] NA
#> 
#> [[6]]$b
#> [1] NA                        "2512-06-17 19:04:57 UTC"
#> [3] NA                        NA                       
#> [5] "2923-03-02 10:49:31 UTC" NA                       
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1] "1892-12-28" "1573-05-22" "2479-12-04" NA           "2112-05-16"
#> [6] "1763-08-02"
#> 
#> [[7]]$b
#> [1]  1  1  1  1  1 NA  1 NA
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1] -440543247
#> 
#> [[8]]$b
#> 01:58:05.432768
#> 02:02:48.023116
#> 
#> 
#> [[9]]
#> [[9]]$a
#>  [1] "1516-07-29" "1055-11-01" "2291-08-28" "1372-06-17" NA          
#>  [6] "2985-11-21" NA           "2675-08-05" "1957-07-24" "1903-03-28"
#> 
#> [[9]]$b
#> 22:52:45.393610
#> 04:13:40.166589
#>              NA
#> 20:01:29.235598
#> 22:28:07.536916
#> 12:24:05.510142
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1]  147307296 -711838904 -734144032
#> 
#> [[10]]$b
#> [1]  368994744         NA  597843959         NA   20763966 -412378578         NA
#> [8]  248486062
#> 
#>