Any list generator

Generate lists containing lists or atomic vectors.

Usage

any_list(len = c(1L, 10L), any_na = FALSE)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10. To produce empty vectors set len = 0L or a range like len = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator object.

Examples

any_list() %>% show_example()
#> [[1]]
#> [[1]]$a
#> 16:28:23.871958
#> 01:44:30.770132
#> 14:47:28.012053
#> 06:37:09.459996
#> 12:54:32.092531
#> 23:46:03.460371
#> 
#> [[1]]$b
#> 23:00:36.630621
#> 05:37:49.507253
#> 13:05:51.521617
#> 13:24:03.273307
#> 20:38:14.157864
#> 
#> 
#> [[2]]
#> [[2]]$a
#> 06:18:31.065891
#> 07:34:22.197663
#> 16:50:56.340640
#> 01:30:42.772177
#> 22:56:39.969526
#> 00:39:37.417303
#> 00:27:09.601931
#> 18:37:21.856317
#> 09:16:48.656631
#> 08:31:21.964256
#> 
#> [[2]]$b
#> 21:31:55.017257
#> 12:32:36.457039
#> 10:08:03.102339
#> 05:30:12.966914
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1] "1493-10-15"
#> 
#> [[3]]$b
#> [1] 0r^a     [4ftej*  8Bi      LSWxPyQU ,k;ujx   0r^a    
#> Levels: 0r^a [4ftej* 8Bi LSWxPyQU ,k;ujx
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] "354-06-30 12:25:20 UTC"  "2416-11-17 22:49:08 UTC"
#> 
#> [[4]]$b
#> [1] "2102-01-28 10:16:22 UTC" "2057-11-24 04:12:34 UTC"
#> [3] "2717-06-05 07:43:22 UTC" "1403-04-27 23:16:19 UTC"
#> [5] "1244-03-17 07:08:42 UTC" "1932-09-25 00:41:01 UTC"
#> [7] "676-06-07 16:41:09 UTC"  "1469-02-02 08:38:51 UTC"
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] tj4"(:2+ V$1      Xl-N|7ze .m$R     d_~^=(3  l3IB]    ?'DY@    YDxB=   
#> Levels: tj4"(:2+ V$1 Xl-N|7ze .m$R d_~^=(3 l3IB] ?'DY@ YDxB=
#> 
#> [[5]]$b
#>  [1] "1901-10-30 11:20:34 UTC" "1016-08-12 00:22:29 UTC"
#>  [3] "13-11-22 11:49:13 UTC"   "1367-12-23 09:32:49 UTC"
#>  [5] "317-06-16 12:41:14 UTC"  "710-07-18 18:55:14 UTC" 
#>  [7] "2639-02-25 15:59:36 UTC" "1494-04-17 20:57:06 UTC"
#>  [9] "984-08-08 04:54:54 UTC"  "53-03-06 01:55:18 UTC"  
#> 
#> 
any_list(len = 10L, any_na = TRUE) %>% show_example()
#> [[1]]
#> [1] NA         NA         "uFcl&J"   NA         "V"        "%w:}zdFH" "Ig"      
#> [8] "jNPyg\"" 
#> 
#> [[2]]
#> [1] "kKW)" " S?" 
#> 
#> [[3]]
#> [1]  472272095 -493131637 -418826291 -337264827 -100771178         NA
#> 
#> [[4]]
#> [1] "C&"       NA         NA         "V=e+!UvK" ".+vQtN"  
#> 
#> [[5]]
#> [1] NA         "0jk"      "#"        NA         "j-"       "]H[h<U?"  "!]|5h["  
#> [8] "rr[]Ozz0"
#> 
#> [[6]]
#>  [1] "{Fr9#D%h" "Ij*<LlDS" NA         NA         "]"        "Gr13.8!x"
#>  [7] "7C`X)Fpi" "JPP{-"    NA         "gU.Q0"   
#> 
#> [[7]]
#> [1] "1922-02-04 08:08:30 UTC" "541-11-28 05:11:47 UTC" 
#> [3] NA                        "1097-07-09 00:49:56 UTC"
#> [5] "2345-11-07 22:07:27 UTC" NA                       
#> [7] "895-03-15 09:51:51 UTC"  "689-11-05 08:49:44 UTC" 
#> 
#> [[8]]
#> [1] "GBe"       "27"        NA          "LSaUv["    "?eHX:[_"   "{#D@%WgXa"
#> [7] NA         
#> 
#> [[9]]
#> [1]  3067 -7692 -6895    NA  2371
#> 
#> [[10]]
#> [1] yc\\x  g9`>gk
#> Levels: yc\\x g9`>gk
#>