Random list generator

Generate random lists containing other lists or atomic vectors of variable lengths.

Usage

any_list(len = c(1L, 10L), frac_na = 0)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_list() |> show_example()
#> [[1]]
#> [[1]]$a
#> [1] FALSE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE
#> 
#> [[1]]$b
#> [1] "ku"     "^> nIW" "]Hi#p"  "D+"    
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1]  TRUE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE
#> 
#> [[2]]$b
#> [1] "2937-06-19" "1009-08-21" "2651-06-19" "2517-12-18" "1053-01-04"
#> [6] "2882-06-23" "1035-11-30" "1332-12-31"
#> 
#> 
#> [[3]]
#> [[3]]$a
#>  [1] "331-10-24 10:57:53 UTC"  "2477-12-15 08:47:32 UTC"
#>  [3] "2296-05-03 02:43:19 UTC" "2792-07-21 09:18:27 UTC"
#>  [5] "1429-03-12 03:55:44 UTC" "121-08-05 07:44:58 UTC" 
#>  [7] "829-08-12 00:08:28 UTC"  "2365-01-28 02:42:05 UTC"
#>  [9] "1231-10-26 16:47:57 UTC" "2198-12-13 19:04:27 UTC"
#> 
#> [[3]]$b
#> [1] "# yFL"     "!U3z}P.=y" "3"         ")JeW"      "JE3]&]\\" 
#> 
#> 
#> [[4]]
#> [[4]]$a
#> 20:58:12.542201
#> 21:01:34.398375
#> 09:31:02.551672
#> 
#> [[4]]$b
#> [1] "!uM]#D"   "/pgCvn"   "TuP'*Kus" "z"        "|K;3P`"  
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] 827605202         0 643071221 783131008         0
#> 
#> [[5]]$b
#> 15:43:42.089260
#> 10:41:30.156877
#> 21:50:58.099697
#> 21:59:32.738726
#> 03:56:07.776733
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1] "1769-05-03 23:03:17 UTC" "113-12-09 17:08:31 UTC" 
#> [3] "2856-12-24 19:27:37 UTC" "738-06-22 21:38:41 UTC" 
#> [5] "834-01-07 14:18:26 UTC"  "176-01-11 05:53:15 UTC" 
#> 
#> [[6]]$b
#> 23:05:16.788562
#> 23:36:02.449283
#> 19:18:19.001172
#> 06:18:38.841659
#> 15:56:28.754783
#> 08:41:43.930247
#> 15:49:45.010449
#> 14:00:28.235422
#> 19:46:36.760384
#> 15:20:42.469943
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1]  2247  2424 -7000
#> 
#> [[7]]$b
#>  [1] "2032-09-28" "1865-05-24" "2228-04-19" "2802-03-18" "1004-03-23"
#>  [6] "1898-09-12" "1286-03-14" "2915-02-08" "1967-12-12" "2012-06-12"
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1] ECL       |ilY[YD   u         Mj o      E         H> k8$*c  .0b.]    
#> [8] 2,tnYZ84, Mj o     
#> Levels: ECL |ilY[YD u Mj o E H> k8$*c .0b.] 2,tnYZ84,
#> 
#> [[8]]$b
#> [1] AI  o   JZ9
#> Levels: AI o JZ9
#> 
#> 
#> [[9]]
#> [[9]]$a
#> [1]  TRUE  TRUE FALSE FALSE FALSE
#> 
#> [[9]]$b
#> [1] -5837 -8954  5697 -1785     0
#> 
#> 
any_list(len = 10L, frac_na = 0.5) |> show_example()
#> [[1]]
#> 12:55:28.672096
#> 
#> [[2]]
#> [1]     NA     NA 201339     NA
#> 
#> [[3]]
#> [1]    0   NA 6555   NA   NA
#> 
#> [[4]]
#> [1] -6239  -463 -5673 -6820
#> 
#> [[5]]
#> [1]  1  1 NA
#> 
#> [[6]]
#> [1]    NA  4410  4411    NA  3462    NA -3535 -1391    NA
#> 
#> [[7]]
#> [1]    0   NA   NA    0 -401   NA    0
#> 
#> [[8]]
#> [1]         NA -449023090
#> 
#> [[9]]
#> [1]      NA -254864 -318038  216451  373844  286467  173978      NA
#> 
#> [[10]]
#> [1] NA  1  1 NA NA  1
#>