Random list generator

Generate random lists containing other lists or atomic vectors of variable lengths.

Usage

any_list(len = c(1L, 10L), frac_na = 0)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_list() |> show_example()
#> [[1]]
#> [[1]]$a
#> [1] "271-12-08 00:02:22 UTC"  "2837-07-13 18:11:28 UTC"
#> [3] "2459-02-12 09:03:43 UTC"
#> 
#> [[1]]$b
#> [1] p8|~Wxh  v3H7/@N.
#> Levels: p8|~Wxh v3H7/@N.
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] "1607-11-11" "1246-05-21" "1523-08-06" "1838-08-21" "1022-06-11"
#> [6] "2224-07-10" "1734-09-05" "2056-02-02"
#> 
#> [[2]]$b
#> 07:34:12.285681
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1] "1038-11-18 09:07:55 UTC" "870-12-11 05:11:47 UTC" 
#> [3] "1897-11-15 17:33:38 UTC"
#> 
#> [[3]]$b
#> 13:36:52.282222
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] "1103-03-28" "2728-02-28"
#> 
#> [[4]]$b
#> 05:03:00.943258
#> 03:07:24.529703
#> 15:36:59.700042
#> 00:53:06.604151
#> 07:29:40.023099
#> 19:36:26.058836
#> 
#> 
#> [[5]]
#> [[5]]$a
#>  [1]          0  175896303  609550854 -231508716  414653376  601386263
#>  [7]  -40868828  715954055  945725629 -165117242
#> 
#> [[5]]$b
#> [1] -731738003 -494821203  538494095
#> 
#> 
#> [[6]]
#> [[6]]$a
#> 02:41:49.468852
#> 10:38:56.501975
#> 11:59:05.188028
#> 03:48:21.420629
#> 17:22:06.529691
#> 09:28:01.304662
#> 03:47:40.069391
#> 12:38:02.819385
#> 00:03:46.167489
#> 14:52:05.819541
#> 
#> [[6]]$b
#> [1] "1452-07-08" "1830-09-26" "2240-09-28"
#> 
#> 
any_list(len = 10L, frac_na = 0.5) |> show_example()
#> [[1]]
#> [[1]]$a
#>              NA
#>              NA
#> 01:45:58.211736
#>              NA
#>              NA
#> 22:49:40.409186
#> 00:57:45.210661
#> 
#> [[1]]$b
#>  [1]    NA  TRUE    NA    NA    NA FALSE    NA    NA    NA    NA
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] p_n8<d
#> Levels: p_n8<d
#> 
#> [[2]]$b
#> [1]      NA -294944  -10969      NA  -50838
#> 
#> 
#> [[3]]
#> [[3]]$a
#>              NA
#> 16:10:17.293445
#>              NA
#> 15:25:49.820502
#> 17:35:53.022320
#>              NA
#> 
#> [[3]]$b
#> [1] NA NA NA
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] "302-11-22 08:48:00 UTC"  NA                       
#> [3] NA                        "2262-11-11 20:30:18 UTC"
#> [5] "669-11-14 03:56:24 UTC"  NA                       
#> 
#> [[4]]$b
#> [1]  TRUE FALSE  TRUE    NA FALSE FALSE FALSE
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] "1771-02-24" NA          
#> 
#> [[5]]$b
#> [1] -20998223        NA
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1]  TRUE  TRUE FALSE
#> 
#> [[6]]$b
#> [1]    NA    NA    NA    NA  3842  8055  1116 -2370
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1]  1 NA
#> 
#> [[7]]$b
#> [1] FALSE FALSE    NA
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1] NA NA
#> 
#> [[8]]$b
#> [1] FALSE    NA  TRUE    NA  TRUE  TRUE    NA    NA  TRUE
#> 
#> 
#> [[9]]
#> [[9]]$a
#> [1] NA NA NA NA  1 NA
#> 
#> [[9]]$b
#> [1] "O"         NA          "Qp8JB"     NA          NA          "G/tHfB7mN"
#> [7] NA         
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1]  0 NA
#> 
#> [[10]]$b
#> [1]          NA          NA          NA 10065609143
#> 
#>