Any list generator

Generate lists containing lists or atomic vectors.

Usage

any_list(len = c(1L, 10L), any_na = FALSE)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10. To produce empty vectors set len = 0L or a range like len = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator object.

Examples

any_list() %>% show_example()
#> [[1]]
#> [1]  813 6805
#> 
#> [[2]]
#> [1] -692948904  840472674
#> 
any_list(len = 10L, any_na = TRUE) %>% show_example()
#> [[1]]
#> [[1]]$a
#> [1] NA          "-Oj\"$b,"  "}\\B1:"    ":R~kv"     "0lHZ&m"    "hV"       
#> [7] "#ltkoLv21" NA         
#> 
#> [[1]]$b
#> [1] "1542-11-14" "1344-06-11" "1670-10-18" "1136-12-02" NA          
#> [6] NA           NA           "1076-08-26"
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] "2082-02-25" "2624-04-17" "2436-02-01" "1392-11-10" NA          
#> [6] "1223-03-07"
#> 
#> [[2]]$b
#> [1] -3433    NA  8981 -2168  1800    NA -2944
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1]  4118 -9240
#> 
#> [[3]]$b
#> [1] NA        NA        "n"       "prx[~dZ" NA       
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1]  102812429  920665378   -7637938  987749376 -212813677         NA          0
#> [8]         NA
#> 
#> [[4]]$b
#> 04:26:31.787732
#> 19:21:18.136671
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] "2751-04-06 02:24:04 UTC" "1844-11-08 00:12:36 UTC"
#> [3] NA                        "507-11-11 23:31:15 UTC" 
#> [5] "1458-03-23 08:31:16 UTC" "1928-01-16 20:18:46 UTC"
#> [7] "2744-03-18 23:47:35 UTC" "2951-07-05 08:30:26 UTC"
#> [9] NA                       
#> 
#> [[5]]$b
#> [1] "899-11-22 04:09:46 UTC"  "2124-01-26 02:32:26 UTC"
#> [3] "1687-02-23 12:26:13 UTC" "834-10-24 14:32:20 UTC" 
#> [5] "1587-08-22 20:51:01 UTC" NA                       
#> [7] "2749-12-31 16:50:52 UTC" "1095-05-24 21:11:24 UTC"
#> [9] "1262-11-03 02:10:04 UTC"
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1]  TRUE  TRUE FALSE FALSE    NA FALSE FALSE FALSE  TRUE
#> 
#> [[6]]$b
#>  [1] -2090  7983  5527  2074 -5798  4712  1760    NA -5351  6313
#> 
#> 
#> [[7]]
#> [[7]]$a
#> 18:27:00.064573
#> 
#> [[7]]$b
#> [1]  232212855  438487912 -739240360  453226068 -290466573 -167634968  795375804
#> [8]          0
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1] NA                        "2635-06-20 16:25:00 UTC"
#> [3] "692-07-17 20:56:57 UTC"  "1250-10-06 21:15:27 UTC"
#> [5] "551-11-13 08:52:54 UTC"  "1667-03-09 14:29:15 UTC"
#> [7] NA                       
#> 
#> [[8]]$b
#> [1]   NA TRUE
#> 
#> 
#> [[9]]
#> [[9]]$a
#> [1]     0 -2613    NA -8834
#> 
#> [[9]]$b
#> [1] "o"       "Z"       "2r"      "@hfCo"   "r"       NA        NA       
#> [8] "6p|<B,F"
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1] "56-03-09 16:52:11 UTC" NA                     
#> 
#> [[10]]$b
#> [1]   NA 3019
#> 
#>