Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator generator

Examples

any_tibble() |> show_example()
#> # A tibble: 8 × 8
#>   ...1       ...2     ...3                       ...4 ...5   ...6  ...7   ...8  
#>   <list>     <fct>    <dttm>                    <dbl> <chr>  <chr> <list> <list>
#> 1 <fct [1]>  RDYdu%n< 2376-07-18 09:57:06  763563722. ";_~_… "?G|" <int>  <date>
#> 2 <chr [10]> }Q+}TL[  2395-02-22 02:46:09 -509319651. "Tf"   "WRD… <date> <date>
#> 3 <lgl [9]>  8M[:X    2876-03-07 08:09:37          0  "NGv}… ",Xg… <time> <date>
#> 4 <time [3]> >        1412-08-29 21:17:18 -440301274. "b+h"  "_2M" <chr>  <date>
#> 5 <lgl [2]>  zGx(,F   1973-07-20 02:02:23 -535264520. " x~Q… "b$@… <time> <date>
#> 6 <lgl [8]>  h]0      126-01-26 08:56:02           0  "@f?t… "89-… <date> <date>
#> 7 <dttm [3]> ^s       1638-11-09 17:04:14  101156195. ";h"   "6eb… <date> <date>
#> 8 <dbl [10]> )VR:Xk8M 1819-07-26 01:33:28 -352823973. "1}Q}… "k.{… <dbl>  <date>
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 × 4
#>    ...1       ...2        ...3       ...4      
#>    <list>     <fct>       <list>     <list>    
#>  1 <fct [7]>  "$R3>"      <dttm [6]> <dttm [1]>
#>  2 <dttm [9]> "IT,Ku=C"   <fct [4]>  <dttm [1]>
#>  3 <dbl [3]>  "9Wh4J"     <dbl [8]>  <dttm [1]>
#>  4 <date [4]> "HU "       <int [7]>  <dttm [1]>
#>  5 <dttm [9]> "t4yHt\"c(" <fct [10]> <dttm [1]>
#>  6 <date [4]> "(6Tvf =C"  <chr [4]>  <dttm [1]>
#>  7 <int [8]>  "["         <chr [2]>  <dttm [1]>
#>  8 <dbl [9]>  "nEDS|CQg&" <lgl [4]>  <dttm [1]>
#>  9 <fct [4]>  "?Fwyn@1yN" <dbl [9]>  <dttm [1]>
#> 10 <fct [8]>  "&1s:g"     <int [7]>  <dttm [1]>
any_tibble(cols = 5L, any_na = TRUE) |> show_example()
#> # A tibble: 8 × 5
#>   ...1             ...2             ...3              ...4 ...5               
#>   <list>           <list>           <date>           <dbl> <dttm>             
#> 1 <named list [2]> <named list [2]> 1807-11-07         NA  1084-03-20 23:25:47
#> 2 <named list [2]> <named list [2]> 1554-11-21 -974571626. 19-10-13 02:25:04  
#> 3 <named list [2]> <named list [2]> 2883-11-12 -608174335. 1960-12-25 02:47:57
#> 4 <named list [2]> <named list [2]> 1820-07-24  129967043. 905-12-29 02:50:36 
#> 5 <named list [2]> <named list [2]> 2415-12-17         NA  1311-01-01 06:55:01
#> 6 <named list [2]> <named list [2]> 2865-02-24  693034606. 2092-07-15 16:27:34
#> 7 <named list [2]> <named list [2]> 1941-07-31  791768456. 471-12-26 12:16:54 
#> 8 <named list [2]> <named list [2]> 1447-10-15 -658565740. 1094-03-23 13:02:34