Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A quickcheck_generator generator

Examples

any_tibble() |> show_example()
#> # A tibble: 7 × 7
#>   ...1                ...2         ...3   ...4                 ...5 ...6        
#>   <dttm>              <list>       <list> <time>              <dbl> <list>      
#> 1 1567-05-17 22:47:47 <named list> <fct>  16:21:41.126822   -7.45e7 <named list>
#> 2 286-12-22 14:49:58  <named list> <chr>  21:46:18.673830    1.68e8 <named list>
#> 3 1135-02-09 23:36:34 <named list> <fct>  04:18:37.823389    9.09e8 <named list>
#> 4 2886-02-03 09:50:47 <named list> <chr>  19:27:26.834145   -4.67e8 <named list>
#> 5 1409-06-12 04:34:24 <named list> <chr>  06:57:47.734160    1.33e8 <named list>
#> 6 1139-07-30 01:00:28 <named list> <fct>  06:00:14.928861    0      <named list>
#> 7 2333-02-21 12:47:43 <named list> <dbl>  10:53:30.599750    5.99e8 <named list>
#> # … with 1 more variable: ...7 <int>
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 × 8
#>    ...1          ...2 ...3         ...4   ...5   ...6   ...7   ...8           
#>    <chr>        <int> <list>       <list> <list> <list> <list> <time>         
#>  1 "$7nV\\UtuD"  8725 <named list> <fct>  <dbl>  <fct>  <fct>  11:38:32.805336
#>  2 "pGqu0o"     -1774 <named list> <fct>  <chr>  <dttm> <fct>  07:09:17.486588
#>  3 "`-"          -570 <named list> <int>  <chr>  <int>  <time> 09:02:46.628176
#>  4 "+zGR_"       6118 <named list> <chr>  <int>  <fct>  <dbl>  09:15:32.893002
#>  5 "86)msn"       762 <named list> <chr>  <chr>  <dbl>  <dttm> 01:09:24.134045
#>  6 "$v/w7 ,Q"   -2058 <named list> <lgl>  <time> <time> <time> 18:27:24.410870
#>  7 "l$0|"           0 <named list> <dbl>  <dttm> <date> <time> 20:29:16.842174
#>  8 "Dt"         -5147 <named list> <dttm> <dttm> <dbl>  <fct>  01:42:46.296209
#>  9 "lbC_~:3d'"   2970 <named list> <date> <chr>  <time> <time> 16:18:10.826957
#> 10 "]e/"         3096 <named list> <date> <dbl>  <lgl>  <dttm> 10:05:52.223272
any_tibble(cols = 5L, frac_na = 0.5) |> show_example()
#> # A tibble: 9 × 5
#>   ...1       ...2        ...3 ...4            ...5     
#>   <list>     <list>     <int> <time>          <list>   
#> 1 <dbl [4]>  <time [1]>    NA 17:20:17.144525 <fct [1]>
#> 2 <date [7]> <time [1]>    NA 07:18:44.170337 <fct [1]>
#> 3 <dttm [2]> <time [1]>  -388 17:58:06.261657 <int [1]>
#> 4 <dbl [3]>  <time [1]>  9614 23:10:53.293242 <int [1]>
#> 5 <dbl [5]>  <time [1]>  6200 05:09:30.438246 <int [1]>
#> 6 <dttm [6]> <time [1]>   943       NA        <int [1]>
#> 7 <dbl [4]>  <time [1]>  6820 20:29:23.196280 <int [1]>
#> 8 <chr [9]>  <time [1]>     0 21:52:50.037022 <int [1]>
#> 9 <int [7]>  <time [1]> -7070 16:21:06.616405 <fct [1]>