Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_tibble() |> show_example()
#> # A tibble: 8 × 9
#>    ...1 ...2     ...3   ...4       ...5    ...6            ...7  ...8           
#>   <int> <list>   <list> <date>     <list>  <time>          <chr> <time>         
#> 1  3912 <named … <date… 1753-09-03 <named… 21:20:29.815346 _A9@… 12:27:58.899415
#> 2 -8853 <named … <dbl … 2878-09-25 <named… 09:55:34.981582 %g    02:04:58.572094
#> 3 -1446 <named … <lgl … 2168-03-30 <named… 05:36:11.661481 eempw 05:51:44.474552
#> 4 -8715 <named … <fct … 2788-03-12 <named… 06:19:10.639389 .LHW? 00:29:42.304023
#> 5     0 <named … <date… 1678-02-22 <named… 07:41:11.909085 s8F)… 06:51:50.057540
#> 6  7756 <named … <time… 1875-11-06 <named… 10:06:06.688868 <a`zJ 07:39:14.523794
#> 7 -5940 <named … <dttm… 2149-04-23 <named… 18:31:26.812785 BsoW… 00:07:17.983342
#> 8   836 <named … <dbl … 2441-04-27 <named… 00:34:08.535707 t6l   21:28:34.562004
#> # … with 1 more variable: ...9 <dbl>
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 × 7
#>    ...1        ...2 ...3             ...4       ...5       ...6          ...7   
#>    <list>     <int> <list>           <list>     <list>     <fct>         <list> 
#>  1 <chr [4]>  -2481 <named list [2]> <int [5]>  <dbl [2]>  "G"           <named…
#>  2 <dbl [8]>      0 <named list [2]> <lgl [8]>  <fct [8]>  "KEV=6V"      <named…
#>  3 <dbl [3]>   7406 <named list [2]> <dbl [1]>  <dbl [3]>  "kfC4u20M"    <named…
#>  4 <lgl [7]>    446 <named list [2]> <fct [5]>  <fct [3]>  "y0L8"        <named…
#>  5 <int [5]>  -1152 <named list [2]> <chr [2]>  <dttm [5]> "w$"          <named…
#>  6 <dttm [9]>  6024 <named list [2]> <time [4]> <fct [6]>  "3{7O:\"Q!\"" <named…
#>  7 <date [2]>  6859 <named list [2]> <date [7]> <chr [4]>  "_zrx"        <named…
#>  8 <chr [10]>  6358 <named list [2]> <dttm [3]> <dbl [1]>  "4?Vs8/"      <named…
#>  9 <time [3]>  2331 <named list [2]> <chr [7]>  <time [7]> "K"           <named…
#> 10 <int [10]>  5975 <named list [2]> <lgl [8]>  <dbl [4]>  "Em ."        <named…
any_tibble(cols = 5L, frac_na = 0.5) |> show_example()
#> # A tibble: 9 × 5
#>    ...1 ...2       ...3              ...4  ...5
#>   <int> <list>     <chr>            <dbl> <int>
#> 1    NA <dttm [2]> NA                 NA      0
#> 2 -9969 <chr [5]>  NA                 NA     NA
#> 3    NA <lgl [8]>  NA        16447997297.    NA
#> 4    NA <dbl [2]>  NA                 NA     NA
#> 5    NA <chr [10]> a@Q9<SZu           NA  -6543
#> 6 -2042 <int [3]>  NA                 NA  -1471
#> 7 -9441 <fct [3]>  t7vE      24945601454.    NA
#> 8    NA <int [7]>  NA       -30116944212.    NA
#> 9     0 <chr [6]>  NA                 NA     NA