Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_tibble() |> show_example()
#> # A tibble: 6 x 2
#>          ...1 ...2      
#>         <dbl> <list>    
#> 1  664958394. <dttm [1]>
#> 2   92190491. <date [1]>
#> 3 -811716403. <dttm [1]>
#> 4   11812892. <lgl [1]> 
#> 5          0  <dbl [1]> 
#> 6          0  <fct [1]> 
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 x 7
#>    ...1              ...2 ...3                 ...4 ...5     ...6      ...7     
#>    <list>           <dbl> <list>              <dbl> <chr>    <list>    <list>   
#>  1 <fct [1]>  -515061806. <named list […    -4.13e8 "Sh`"    <lgl [7]> <date [1…
#>  2 <lgl [1]>   -92773989. <named list […    -2.71e8 "R08HQ!" <time [7… <int [1]>
#>  3 <chr [1]>   371899164. <named list […    -2.01e8 "+&"     <date [1… <dbl [1]>
#>  4 <date [1]> -922237888. <named list […    -8.16e8 "].?9H"  <dttm [6… <fct [1]>
#>  5 <lgl [1]>   265364392. <named list […     1.91e8 "/A"     <lgl [4]> <dbl [1]>
#>  6 <chr [1]>  -955416591. <named list […     3.56e8 "`Tp5Y"  <lgl [2]> <dttm [1…
#>  7 <time [1]>  387956164. <named list […     4.76e8 "e"      <fct [2]> <dbl [1]>
#>  8 <dbl [1]>           0  <named list […     6.03e8 "f,\\F"  <time [1… <dttm [1…
#>  9 <chr [1]>  -228044787. <named list […    -9.47e8 "J5~vnr… <dttm [3… <fct [1]>
#> 10 <chr [1]>           0  <named list […    -7.72e8 "ftH"    <time [6… <chr [1]>
any_tibble(cols = 5L, frac_na = 0.5) |> show_example()
#> # A tibble: 10 x 5
#>     ...1 ...2       ...3                 ...4 ...5      
#>    <int> <list>     <dttm>              <int> <list>    
#>  1    NA <chr [1]>  2855-08-31 23:37:08    NA <fct [1]> 
#>  2    NA <date [1]> 2089-01-09 07:20:09    NA <date [1]>
#>  3    NA <lgl [1]>  NA                     NA <chr [1]> 
#>  4     1 <dbl [1]>  1243-03-19 10:47:35    NA <time [1]>
#>  5    NA <chr [1]>  0708-10-04 10:43:32     1 <chr [1]> 
#>  6    NA <lgl [1]>  1611-08-29 19:12:43     1 <dbl [1]> 
#>  7    NA <chr [1]>  NA                      1 <lgl [1]> 
#>  8    NA <dttm [1]> 0831-11-22 11:30:44    NA <int [1]> 
#>  9     1 <dbl [1]>  0668-01-12 21:04:14    NA <int [1]> 
#> 10     1 <int [1]>  1320-01-18 11:11:25    NA <int [1]>