Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A quickcheck_generator generator

Examples

any_tibble() |> show_example()
#> # A tibble: 9 × 9
#>   ...1        ...2       ...3      ...4   ...5        ...6 ...7     ...8    ...9
#>   <list>      <date>     <list>    <list> <list>     <dbl> <fct>   <dbl>   <dbl>
#> 1 <dttm [2]>  2828-08-17 <dbl [1]> <dttm… <named…   2.32e8 "]iO…  7.98e8 -5.46e8
#> 2 <chr [6]>   1966-07-01 <dbl [1]> <time… <named…   4.38e8 "neQ…  8.65e8  6.89e8
#> 3 <date [1]>  2302-12-21 <dbl [1]> <dbl … <named…   8.27e8 "),N… -1.99e8 -2.36e8
#> 4 <dttm [1]>  2147-07-13 <dbl [1]> <int … <named…   8.30e8 "3aD…  0      -3.04e8
#> 5 <date [10]> 2949-12-22 <dbl [1]> <dbl … <named…   5.49e8 "G%v…  0       7.92e8
#> 6 <dttm [3]>  1540-12-26 <dbl [1]> <date… <named…  -9.97e8 "9I"  -7.14e8  7.99e8
#> 7 <chr [6]>   1580-08-27 <dbl [1]> <dbl … <named…  -8.07e7 "ci#…  0       6.98e8
#> 8 <lgl [5]>   1123-07-23 <dbl [1]> <time… <named…   9.49e8 "l"    3.62e8 -3.98e8
#> 9 <fct [5]>   2696-01-08 <dbl [1]> <chr … <named…   0      "BVF…  3.81e8 -7.50e8
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 × 6
#>    ...1       ...2            ...3    ...4   ...5            ...6               
#>    <list>     <time>          <list>  <fct>  <time>          <dttm>             
#>  1 <chr [1]>  22:21:51.067013 <int [… ".>kj… 16:02:31.314402 284-02-20 16:02:01 
#>  2 <dbl [1]>  19:51:12.761064 <int [… "Stcp… 07:18:38.622875 2846-06-01 15:46:30
#>  3 <lgl [1]>  05:40:33.402656 <fct [… "DqP"  22:16:35.212334 2612-04-14 21:39:59
#>  4 <lgl [1]>  21:58:28.835628 <dttm … "^wqB… 17:11:44.271087 2292-05-26 03:43:57
#>  5 <dttm [1]> 14:49:45.602452 <date … "cx"   11:09:28.062292 2165-12-04 14:51:09
#>  6 <time [1]> 21:10:21.993531 <chr [… "p"    03:25:02.709342 301-08-30 14:28:03 
#>  7 <date [1]> 05:45:19.443057 <lgl [… "\"V"  21:52:21.629559 1072-12-30 15:40:34
#>  8 <time [1]> 01:09:47.805962 <fct [… "nS[W… 19:26:45.900844 395-07-31 17:58:35 
#>  9 <time [1]> 20:08:30.418883 <dttm … ")Ii7… 17:49:51.391513 1499-12-12 06:24:33
#> 10 <time [1]> 11:23:12.711360 <date … "6Z*_" 01:29:49.365501 29-06-09 15:11:04  
any_tibble(cols = 5L, frac_na = 0.5) |> show_example()
#> # A tibble: 8 × 5
#>   ...1        ...2 ...3       ...4       ...5    
#>   <list>     <int> <date>     <list>     <chr>   
#> 1 <dbl [9]>     NA 2528-11-27 <dbl [3]>  NA      
#> 2 <int [6]>      1 NA         <time [5]> a(X*PL  
#> 3 <date [2]>    NA 1954-01-01 <dttm [7]> HUz!$)EC
#> 4 <time [1]>    NA NA         <dbl [6]>  NA      
#> 5 <int [10]>     1 2946-09-08 <int [3]>  <{NQr}S 
#> 6 <time [9]>    NA 1580-05-03 <int [9]>  NA      
#> 7 <time [5]>    NA NA         <time [5]> NA      
#> 8 <time [8]>     1 1009-05-30 <chr [3]>  NA