Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A quickcheck_generator generator

Examples

any_tibble() |> show_example()
#> # A tibble: 6 × 9
#>   ...1            ...2     ...3    ...4  ...5    ...6   ...7    ...8  ...9      
#>   <time>          <list>   <list>  <lgl> <list>  <fct>  <list>  <fct> <date>    
#> 1 05:39:18.068190 <named … <fct [… TRUE  <dttm … "< "   <named… ;AJ4… 2491-02-25
#> 2 17:26:18.094193 <named … <fct [… TRUE  <dttm … "^sA$… <named… <YK|… 2300-07-13
#> 3 01:39:48.096100 <named … <fct [… FALSE <dttm … "NQ"   <named… gHIj… 2352-06-15
#> 4 16:13:24.703899 <named … <fct [… FALSE <dttm … "*E"   <named… /uK8… 1991-08-23
#> 5 05:46:05.810102 <named … <fct [… FALSE <dttm … "}"    <named… ng(k… 2368-02-01
#> 6 15:48:45.435240 <named … <fct [… TRUE  <dttm … "~gbA… <named… z=    2873-05-17
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 × 7
#>    ...1         ...2       ...3        ...4                ...5    ...6   ...7  
#>    <chr>        <list>     <list>      <dttm>              <list>  <list> <list>
#>  1 "L\\O}},R"   <fct [3]>  <dttm [10]> 1723-09-04 22:01:06 <lgl [… <chr … <chr …
#>  2 "Kp>f"       <date [5]> <int [1]>   1405-04-30 00:22:19 <time … <chr … <time…
#>  3 "jt]V"       <lgl [10]> <time [7]>  697-06-15 05:21:15  <dttm … <date… <time…
#>  4 "!2CCv"      <chr [8]>  <date [1]>  1381-05-11 12:32:47 <dbl [… <lgl … <time…
#>  5 "Pf>_p"      <fct [8]>  <time [6]>  1708-01-22 03:49:33 <lgl [… <lgl … <date…
#>  6 "9qLcB=gnc"  <int [2]>  <chr [8]>   1414-04-16 03:53:48 <date … <chr … <time…
#>  7 "q~b$Fg??"   <chr [3]>  <int [10]>  1783-07-31 15:25:32 <date … <fct … <int …
#>  8 "R4x"        <dbl [10]> <int [1]>   1508-03-18 10:17:50 <date … <int … <dbl …
#>  9 "M!b\"X+H{/" <int [1]>  <time [5]>  2660-10-26 04:04:20 <int [… <fct … <fct …
#> 10 "@G#"        <date [2]> <fct [2]>   1181-09-09 01:26:43 <lgl [… <int … <int …
any_tibble(cols = 5L, frac_na = 0.5) |> show_example()
#> # A tibble: 5 × 5
#>   ...1      ...2              ...3 ...4       ...5            
#>   <list>    <list>           <dbl> <list>     <list>          
#> 1 <chr [1]> <dttm [1]> -310031251. <lgl [1]>  <named list [2]>
#> 2 <chr [1]> <chr [1]>          NA  <date [1]> <named list [2]>
#> 3 <chr [1]> <lgl [1]>          NA  <lgl [1]>  <named list [2]>
#> 4 <chr [1]> <dbl [1]>          NA  <lgl [1]>  <named list [2]>
#> 5 <chr [1]> <dttm [1]> -128157782. <time [1]> <named list [2]>