Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_tibble() |> show_example()
#> # A tibble: 7 × 4
#>   ...1                ...2     ...3            ...4      
#>   <dttm>              <chr>    <time>          <date>    
#> 1 2573-04-26 08:35:22 "5"      15:03:19.515010 2469-01-30
#> 2 1810-10-13 05:56:15 "q"      13:57:54.571991 1314-12-20
#> 3 1462-07-23 06:39:32 "SC"     07:06:59.406497 1191-05-31
#> 4 2011-12-02 03:27:54 "m9Uto"  05:33:50.402426 2356-03-05
#> 5 785-06-28 01:13:00  "\"Fg*"  00:31:13.179626 2836-09-06
#> 6 2437-08-03 21:06:05 "=VT!|<" 18:25:27.646369 1226-10-01
#> 7 1361-04-23 13:44:36 "#@"     04:06:12.289330 2016-09-30
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 × 6
#>    ...1                    ...2 ...3       ...4       ...5       ...6 
#>    <list>                 <dbl> <list>     <list>     <list>     <lgl>
#>  1 <named list [2]>  773911305. <dttm [3]> <fct [6]>  <dbl [1]>  FALSE
#>  2 <named list [2]>          0  <lgl [4]>  <chr [4]>  <int [1]>  FALSE
#>  3 <named list [2]>  910273841. <lgl [3]>  <date [3]> <time [1]> FALSE
#>  4 <named list [2]>    7455515. <chr [4]>  <fct [4]>  <dbl [1]>  TRUE 
#>  5 <named list [2]>  374939787. <chr [1]>  <dttm [5]> <int [1]>  TRUE 
#>  6 <named list [2]> -455743587. <int [1]>  <chr [9]>  <dbl [1]>  FALSE
#>  7 <named list [2]> -274761078. <date [7]> <int [8]>  <dbl [1]>  TRUE 
#>  8 <named list [2]>  123758521. <dbl [2]>  <chr [5]>  <dttm [1]> TRUE 
#>  9 <named list [2]>  363660587. <dbl [3]>  <chr [8]>  <dbl [1]>  TRUE 
#> 10 <named list [2]> -999619017. <dbl [3]>  <time [8]> <dbl [1]>  TRUE 
any_tibble(cols = 5L, frac_na = 0.5) |> show_example()
#> # A tibble: 10 × 5
#>     ...1 ...2                ...3 ...4             ...5      
#>    <int> <list>             <dbl> <list>           <list>    
#>  1    NA <named list [2]>      NA <named list [2]> <dbl [1]> 
#>  2    NA <named list [2]>      NA <named list [2]> <chr [1]> 
#>  3    NA <named list [2]>   68413 <named list [2]> <dbl [1]> 
#>  4    NA <named list [2]>   32194 <named list [2]> <int [1]> 
#>  5    NA <named list [2]> -323401 <named list [2]> <time [1]>
#>  6    NA <named list [2]> -202443 <named list [2]> <chr [1]> 
#>  7     1 <named list [2]>      NA <named list [2]> <int [1]> 
#>  8     1 <named list [2]>   91582 <named list [2]> <int [1]> 
#>  9     1 <named list [2]>  230187 <named list [2]> <int [1]> 
#> 10    NA <named list [2]>      NA <named list [2]> <time [1]>