Random tibble generator

Generate random tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), frac_na = 0)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A quickcheck_generator generator

Examples

any_tibble() |> show_example()
#> # A tibble: 9 × 5
#>    ...1 ...2       ...3  ...4         ...5
#>   <int> <list>     <lgl> <list>      <int>
#> 1  6925 <dttm [1]> FALSE <fct [9]>   -8974
#> 2  9513 <dttm [1]> FALSE <time [10]>     0
#> 3  7485 <dttm [1]> TRUE  <int [1]>    8561
#> 4 -6169 <dttm [1]> FALSE <chr [5]>   -7728
#> 5  3236 <dttm [1]> TRUE  <date [2]>   -360
#> 6  7215 <dttm [1]> FALSE <lgl [4]>       0
#> 7   334 <dttm [1]> TRUE  <dttm [8]>      0
#> 8  8096 <dttm [1]> TRUE  <lgl [10]>   6310
#> 9 -3768 <dttm [1]> TRUE  <chr [2]>       0
any_tibble(rows = 10L) |> show_example()
#> # A tibble: 10 × 4
#>    ...1                ...2                       ...3 ...4      
#>    <dttm>              <dttm>                    <dbl> <date>    
#>  1 1854-03-18 03:33:40 2043-05-22 03:47:22  895492233. 2203-05-12
#>  2 1353-09-02 00:43:04 568-03-11 15:36:37   612191975. 2006-07-15
#>  3 1495-11-17 04:44:04 2802-08-17 07:08:43          0  2138-02-16
#>  4 1608-09-17 20:12:33 2897-05-24 17:06:12 -889012890. 2890-04-01
#>  5 482-06-02 19:41:54  1987-06-08 16:39:35 -891681704. 2499-12-11
#>  6 2122-05-08 10:58:58 537-08-31 07:04:07           0  1344-06-07
#>  7 962-07-23 10:38:00  2560-01-25 16:04:31 -805017660. 1490-09-23
#>  8 2908-12-29 20:32:50 1764-06-24 14:36:01 -285345415. 2168-09-27
#>  9 1530-12-19 03:42:06 561-11-13 22:18:33   -16350517. 2811-04-17
#> 10 2322-04-27 22:33:53 1057-02-28 03:07:46  566202283. 2753-07-15
any_tibble(cols = 5L, frac_na = 0.5) |> show_example()
#> # A tibble: 2 × 5
#>   ...1       ...2  ...3 ...4       ...5     
#>   <list>    <dbl> <int> <list>     <list>   
#> 1 <dbl [1]>    NA     1 <date [2]> <lgl [6]>
#> 2 <dbl [1]>    NA    NA <fct [2]>  <chr [7]>