Any tibble generator

Generate tibbles.

Usage

any_tibble(rows = c(1L, 10L), cols = c(1L, 10L), any_na = FALSE)

Arguments

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

any_na

Whether NA values should be allowed.

Value

A quickcheck_generator object.

Examples

any_tibble(rows = 3L, cols = 3L) %>% show_example()
#> # A tibble: 3 × 3
#>   ...1       ...2       ...3            
#>   <date>     <date>     <list>          
#> 1 2509-02-11 1557-12-16 <named list [2]>
#> 2 1786-10-05 1729-08-03 <named list [2]>
#> 3 1300-02-08 1729-10-26 <named list [2]>