Any vector generator

Generate random atomic vectors or lists.

Usage

any_vector(len = 1L, frac_na = 0)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10.

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A hedgehog.internal.gen generator

Examples

any_vector() |> show_example()
#> [1] -85751
any_vector(len = 10L, frac_na = 0.5) |> show_example()
#> [[1]]
#> [[1]]$a
#> [1] <NA>
#> Levels: 
#> 
#> [[1]]$b
#> [1] <NA> <NA> p    <NA>
#> Levels: p
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1]      NA -680856      NA
#> 
#> [[2]]$b
#> [1]        NA        NA 963736768 479594775
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1]         NA         NA -879880998 -302448999   85531510  764796018 -897217131
#> [8]         NA   54434912
#> 
#> [[3]]$b
#> [1] "1842-06-10" "0726-01-01"
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1]           NA -21109723170  26141979917
#> 
#> [[4]]$b
#> [1] "0840-04-29 18:24:32 LMT" "0623-05-28 19:43:27 LMT"
#> [3] NA                        "2988-01-14 00:54:36 PST"
#> [5] NA                        NA                       
#> 
#> 
#> [[5]]
#> [[5]]$a
#>  [1] "0885-02-02" NA           "0365-09-16" "0437-05-06" "0550-02-19"
#>  [6] "1442-09-23" NA           NA           NA           "0596-07-15"
#> 
#> [[5]]$b
#> [1]  TRUE    NA    NA    NA FALSE  TRUE  TRUE    NA
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1] NA  NA  "C" "&" NA 
#> 
#> [[6]]$b
#> [1] "0398-04-24 15:03:09 LMT" NA                       
#> [3] "1639-04-24 09:06:59 LMT" "2091-06-30 12:12:15 PDT"
#> [5] NA                       
#> 
#> 
#> [[7]]
#> [[7]]$a
#>  [1]           NA           NA -15011400047           NA           NA
#>  [6]           NA -49174859461 -38940301991           NA           NA
#> 
#> [[7]]$b
#> [1]           NA           NA           NA           NA           NA
#> [6]           NA           NA -45891135264           NA
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1]         NA         NA         NA         NA  194607184 -135749339
#> 
#> [[8]]$b
#> [1] "." "L" "9" NA 
#> 
#> 
#> [[9]]
#> [[9]]$a
#> [1]       NA -5148629
#> 
#> [[9]]$b
#> [1] "0212-11-25 11:46:27 LMT" "2933-09-12 21:23:32 PDT"
#> [3] "1216-09-23 02:14:32 LMT" NA                       
#> [5] "0827-08-12 17:09:36 LMT" NA                       
#> [7] NA                       
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1] NA
#> 
#> [[10]]$b
#> [1] "0470-03-12 20:59:56 LMT" NA                       
#> [3] "0209-09-04 18:15:53 LMT" NA                       
#> [5] NA                        NA                       
#> 
#>