Random vector generator

Generate random atomic vectors or lists.

Usage

any_vector(len = c(1L, 10L), frac_na = 0)

Arguments

len

Length of the generated vectors. If len is a single number all vectors will have this length. If len is a numeric vector of length 2 it will produce vectors with lengths between a minimum and maximum, inclusive. For example len = c(1L, 10L) would produce vectors with lengths between 1 and 10. To produce empty vectors set len = 0L or a range like len = c(0L, 10L).

frac_na

The approximate fraction of NA values to be included in the vector. A fraction of 0, the default, will include no NA values. A fraction of 0.5 would include around 50% NA values.

Value

A quickcheck_generator generator

Examples

any_vector() |> show_example()
#> [[1]]
#> [[1]]$a
#> [1] -442994332 -229608035 -908390759 -952270947 -437643923
#> 
#> [[1]]$b
#>  [1] *r        !t|D`     ydZ*v     ]4G*eE1Qi )W'9zfGr& )^A|I/+   \\VE)0Yff
#>  [8] v$|?w     &z        ycI      
#> 10 Levels: *r !t|D` ydZ*v ]4G*eE1Qi )W'9zfGr& )^A|I/+ \\VE)0Yff v$|?w ... ycI
#> 
#> 
#> [[2]]
#> [[2]]$a
#> [1] "757-11-10 05:22:40 UTC" "948-05-21 19:24:00 UTC" "563-09-11 20:51:15 UTC"
#> 
#> [[2]]$b
#> [1]   429  2903 -3053  2980 -8536 -9676  8415     0
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1] U|Y@yT   !UNh     h}[P0NBT nc       '`0&v6   Q$o     
#> Levels: U|Y@yT !UNh h}[P0NBT nc '`0&v6 Q$o
#> 
#> [[3]]$b
#> [1] FALSE  TRUE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] "831-01-08 11:58:22 UTC"  "2968-10-17 23:55:22 UTC"
#> [3] "1566-02-01 07:48:52 UTC" "1980-05-19 11:26:25 UTC"
#> [5] "586-02-27 00:56:14 UTC"  "2808-04-14 14:41:05 UTC"
#> [7] "834-07-09 10:15:55 UTC"  "2047-08-04 09:19:35 UTC"
#> [9] "199-09-28 02:42:20 UTC" 
#> 
#> [[4]]$b
#> [1] "2954-10-19 06:20:06 UTC" "1374-10-28 19:30:06 UTC"
#> [3] "2682-09-03 23:44:16 UTC"
#> 
#> 
#> [[5]]
#> [[5]]$a
#> [1] FALSE FALSE FALSE
#> 
#> [[5]]$b
#> [1] "2150-08-03 00:58:29 UTC" "221-01-26 22:09:19 UTC" 
#> [3] "1554-07-01 03:26:41 UTC" "1544-04-01 02:50:29 UTC"
#> [5] "1357-01-18 13:41:23 UTC" "2834-02-17 07:53:18 UTC"
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1] "a"         "s]#p@&0\"" "N"         "*l)\"<?I|" "<ZZP9"     "{"        
#> 
#> [[6]]$b
#> [1] YjR        v0W`\\)~A@ 0$EZsL:3^  \\"$OV     WQa4Y      sV]wt      5GFw>f    
#> Levels: YjR v0W`\\)~A@ 0$EZsL:3^ \\"$OV WQa4Y sV]wt 5GFw>f
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1] 814932105 845995541 195540267 469913034 414657607
#> 
#> [[7]]$b
#> [1] FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE  TRUE  TRUE
#> 
#> 
#> [[8]]
#> [[8]]$a
#> 19:06:37.836165
#> 12:39:25.695204
#> 14:20:08.747924
#> 10:44:08.386876
#> 02:51:24.355439
#> 
#> [[8]]$b
#> [1]  TRUE  TRUE  TRUE FALSE  TRUE
#> 
#> 
#> [[9]]
#> [[9]]$a
#> [1] FALSE  TRUE  TRUE
#> 
#> [[9]]$b
#> [1] "1620-01-07" "1578-09-17" "2219-02-26" "2901-07-02"
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1] 5859 2556 7677 5508    0 5949
#> 
#> [[10]]$b
#> 10:24:09.634672
#> 20:13:02.257533
#> 08:19:12.102170
#> 09:24:17.244641
#> 11:45:54.101455
#> 21:55:01.080942
#> 16:43:19.789459
#> 08:16:00.861043
#> 
#> 
any_vector(len = 10L, frac_na = 0.5) |> show_example()
#> [[1]]
#> [[1]]$a
#>  [1]    NA  1246    NA    NA -2121  2963 -8714    NA    NA -1636
#> 
#> [[1]]$b
#> [1]  1  1 NA  1 NA NA
#> 
#> 
#> [[2]]
#> [[2]]$a
#>  [1]         NA  765594163  930620327         NA         NA         NA
#>  [7]  402511770 -886119782         NA         NA
#> 
#> [[2]]$b
#> [1]    NA    NA  TRUE FALSE  TRUE FALSE    NA
#> 
#> 
#> [[3]]
#> [[3]]$a
#> [1] "W&]!"
#> 
#> [[3]]$b
#> [1] -8479  7997 -3449  2491    NA    NA    NA
#> 
#> 
#> [[4]]
#> [[4]]$a
#> [1] NA
#> 
#> [[4]]$b
#> [1]  1  1  1  1 NA  1
#> 
#> 
#> [[5]]
#> [[5]]$a
#> NA
#> NA
#> NA
#> NA
#> 
#> [[5]]$b
#> [1] -256741264         NA         NA         NA
#> 
#> 
#> [[6]]
#> [[6]]$a
#> [1] "q$={9^"  NA        NA        NA        "<7ovADJ" "yUEn"    "Hf#'"   
#> [8] "3"      
#> 
#> [[6]]$b
#> [1]   NA TRUE TRUE
#> 
#> 
#> [[7]]
#> [[7]]$a
#> [1] FALSE    NA
#> 
#> [[7]]$b
#> [1]  1 NA  1  1 NA NA NA NA  1
#> 
#> 
#> [[8]]
#> [[8]]$a
#> [1] FALSE    NA    NA FALSE
#> 
#> [[8]]$b
#> [1]      NA -212893      NA  -76819 -141961      NA   76133
#> 
#> 
#> [[9]]
#> [[9]]$a
#> [1]  1 NA  1 NA NA NA
#> 
#> [[9]]$b
#> [1]  10181329 559319823        NA 634470065        NA 713727660
#> 
#> 
#> [[10]]
#> [[10]]$a
#> [1] "1998-08-08" NA           NA           NA           "1022-02-05"
#> [6] NA           "1318-06-11"
#> 
#> [[10]]$b
#> [1] "900-05-08 08:43:20 UTC"  NA                       
#> [3] NA                        NA                       
#> [5] NA                        "2106-05-31 10:55:21 UTC"
#> [7] NA                       
#> 
#>