Any R object generator

Generate any R object. This doesn't actually generate any possible object, just the most common ones, namely atomic vectors, lists, data.frames, tibbles, data.tables, and undefined values like NA, NULL, Inf, and NaN.

Usage

anything(any_empty = TRUE, any_undefined = TRUE)

Arguments

any_empty

Whether empty vectors or data frames should be allowed.

any_undefined

Whether undefined values should be allowed.

Value

A quickcheck_generator object.

Examples

anything() |> show_example()
#>                   ...1      ...2 ...3
#> 1: 1621-06-19 12:07:54 <list[2]>    1
#> 2: 2505-02-23 03:48:37 <list[2]>   NA
#> 3:  835-11-17 23:47:41 <list[2]>    1
#> 4: 2389-05-25 19:03:05 <list[2]>    1
#> 5: 1335-11-29 01:29:44 <list[2]>    1
#> 6: 2499-01-19 02:45:27 <list[2]>    1
#> 7: 2748-02-09 10:26:22 <list[2]>    1
#> 8:                <NA> <list[2]>    1
#>                                                                 ...4   ...5
#> 1:                                   |&,favWv8,gC,NA,lnibS5,aP+:,...   EfIF
#> 2:                                                     1, 1,NA, 1,NA     ])
#> 3:                               NA,O|eCM~%n,[3cCs,!c,NA,jMt%cAG,...    skW
#> 4:                         F{;0|,t B/$W8|b,g]ec0{,^a<m [y.Z,yS,%,...  13fHw
#> 5: -678285864,-723922690, 727324165, 873772130,        NA,        NA    ~QQ
#> 6:                                      TRUE,   NA,FALSE,FALSE, TRUE `f=:3`
#> 7:                                 TRUE,  NA,TRUE,  NA,TRUE,TRUE,...     iW
#> 8:                                              NA,   NA,   NA,FALSE      A
#>               ...6                ...7
#> 1: 12:05:33.162926 1401-11-09 22:12:11
#> 2: 16:45:24.735695                  x6
#> 3: 22:49:34.347370 1541-04-12 12:37:18
#> 4: 11:54:34.755047          2683-02-21
#> 5: 18:41:57.082659            (exNb G3
#> 6:              NA           750913271
#> 7: 04:05:48.004147                  NA
#> 8: 19:00:31.145194              )Jj,V~