Any R object generator

Generate any R object. This doesn't actually generate any possible object, just the most common ones, namely atomic vectors, lists, data.frames, tibbles, data.tables, and undefined values like NA, NULL, Inf, and NaN.

Usage

anything(any_empty = TRUE, any_undefined = TRUE)

Arguments

any_empty

Whether empty vectors or data frames should be allowed.

any_undefined

Whether undefined values should be allowed.

Value

A quickcheck_generator object.

Examples

anything() |> show_example()
#>                  ...1      ...2     ...3  ...4            ...5       ...6
#> 1 2020-11-25 00:43:58  BWC$)nid    O!v56 FALSE 12:12:39.958327       <NA>
#> 2 1196-05-03 13:56:46      ,_Vq     <NA> FALSE 07:56:01.833668       <NA>
#> 3                <NA>     3 k^d  Fv".]12    NA 21:42:03.193964 2627-09-17
#> 4  486-12-25 21:00:50     HgSc>     <NA> FALSE 16:46:56.927288 1267-09-14
#> 5 2191-05-12 03:46:29         [ '0@8%f$H  TRUE 19:01:28.269671 1620-09-29
#> 6 1404-09-01 18:38:33 ^DfQO.dmq      A{p  TRUE 14:00:23.838373 2246-05-07
#> 7                <NA>    s{rR^f    zZ'+6  TRUE              NA 2736-08-18
#> 8                <NA>      <NA>     <NA>    NA 13:44:41.086967 1053-01-30
#> 9 1477-12-02 10:48:13    ~MAVr'   Zgv5]m FALSE 23:47:35.344948 2002-10-26
anything(any_empty = FALSE, any_undefined = FALSE) |> show_example()
#>                  ...1
#> 1 1179-10-13 10:14:33
#> 2 2519-09-30 10:40:26
#> 3 1939-06-16 02:42:11
#> 4 1453-01-27 21:14:01
#> 5    6-01-24 08:30:40
#> 6 2632-06-29 08:02:50