Random data.table generator

Random data.table generator

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

data.table_of(logical_(), date_()) |> show_example()
#>     ...1  ...2       ...3       ...4  ...5  ...6
#> 1: FALSE FALSE 2594-12-27 2864-01-07  TRUE  TRUE
#> 2: FALSE  TRUE 1283-02-17 2418-02-17  TRUE FALSE
#> 3:  TRUE FALSE 1102-09-18 2375-09-16 FALSE FALSE
data.table_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>                    ...1       ...2     ...3       ...4       ...5
#>  1: 2380-06-01 18:20:17 1782-05-08 *,jw~IKm -873028201   19065626
#>  2: 2443-10-09 12:45:29 2387-03-18      x+Q -439491223 -562755985
#>  3:  860-10-18 17:01:36 1272-07-18 i$kO~q"P  -55438169  639138054
#>  4: 2347-02-16 20:13:30 2924-12-10  RU#f$9K  537946860 -686221601
#>  5:  620-07-23 03:18:17 2361-08-10       %S  872829799  835788447
#>  6:  443-04-23 01:19:12 1833-06-20   bG:VW]  159544012  959548898
#>  7: 1697-11-13 11:36:11 2452-06-23    NQo,K -147893243   13120284
#>  8: 2295-02-27 00:40:44 1390-09-30     Z)tF          0 -803081909
#>  9: 1853-02-09 00:42:26 1470-01-01 h'<Agy8'  534437982  652891662
#> 10:  392-07-25 13:59:07 2995-04-19    ;F+hE  713147601  250907784