data.table generator with randomized columns

data.table generator with randomized columns

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data.table_of(logical_(), date_()) |> show_example()
#>          ...1  ...2  ...3       ...4  ...5       ...6  ...7
#> 1: 2686-03-05 FALSE FALSE 1933-08-08 FALSE 1682-03-04  TRUE
#> 2: 1383-07-13  TRUE  TRUE 1954-10-20 FALSE 2459-03-15  TRUE
#> 3: 1261-12-19  TRUE  TRUE 2530-11-22 FALSE 2843-07-21 FALSE
#> 4: 1310-06-09 FALSE  TRUE 1295-06-20  TRUE 2186-12-21  TRUE
#> 5: 2715-05-15  TRUE FALSE 2597-07-20 FALSE 1236-06-30  TRUE
#> 6: 1161-12-24 FALSE FALSE 2410-09-17  TRUE 1448-07-04  TRUE
#> 7: 2592-09-28 FALSE  TRUE 2385-10-09  TRUE 2741-09-08 FALSE
data.table_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>           ...1       ...2            ...3     ...4  ...5
#>  1: 2142-02-02 1145-07-15 02:35:05.338727        h  TRUE
#>  2: 1451-02-19 1711-10-17 06:03:00.916826   R)`UY# FALSE
#>  3: 2856-03-05 2261-03-17 18:54:11.624729     \\#M  TRUE
#>  4: 1293-03-15 1367-07-28 10:58:22.287127      .Q%  TRUE
#>  5: 1317-05-27 2110-08-25 23:29:08.096178      QkE FALSE
#>  6: 2306-10-02 2923-11-04 02:56:30.982373    +NWh'  TRUE
#>  7: 1955-08-16 1197-03-24 20:20:41.705104       (z FALSE
#>  8: 1368-05-29 2218-12-28 05:24:52.418068 ;j,;%r{! FALSE
#>  9: 1977-05-08 1275-06-12 07:17:36.746920      I&5  TRUE
#> 10: 1962-11-18 1116-10-02 10:24:05.667771      Wgk  TRUE