data.table generator with randomized columns

data.table generator with randomized columns

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data.table_of(logical_(), date_()) |> show_example()
#>          ...1       ...2 ...3
#> 1: 1559-02-01 1414-09-04 TRUE
#> 2: 1876-03-08 2426-03-17 TRUE
#> 3: 1507-09-07 2899-12-10 TRUE
data.table_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>           ...1            ...2      ...3            ...4                ...5
#>  1:  982441839 12:46:55.252479    x1-th" 23:55:57.518750 2634-02-10 09:28:15
#>  2:  891279215 10:15:43.399343 20dnt+80L 03:33:12.981314 2712-05-16 16:58:15
#>  3:   61605954 03:16:29.258826      %RhB 19:30:26.823904   31-06-26 19:00:47
#>  4: -486356954 09:58:18.178184      _BN/ 23:58:25.353553 2716-03-17 17:36:11
#>  5:          0 23:54:20.436956      "TNF 04:40:35.107253 2205-12-19 17:25:33
#>  6:  356033941 15:03:51.010411  ?qtAGN-e 09:26:13.495873 2365-07-09 00:56:24
#>  7:  469012778 12:35:54.282121  ^#fYJ`?v 00:22:51.489642 2551-12-31 22:09:53
#>  8:  -38190810 10:47:13.759404        Yz 17:29:06.453774 2854-04-24 17:59:07
#>  9: -329302224 23:39:56.201395  ml(0` e@ 11:02:30.914887 1000-11-30 09:54:27
#> 10: -905648463 06:01:56.705807     %Js u 01:11:33.999671 2000-10-01 07:52:16