Random data.table generator

Random data.table generator

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

data.table_of(logical_(), date_()) |> show_example()
#>     ...1  ...2  ...3
#> 1:  TRUE FALSE FALSE
#> 2: FALSE FALSE FALSE
#> 3: FALSE FALSE FALSE
#> 4:  TRUE  TRUE  TRUE
data.table_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>      ...1       ...2            ...3       ...4  ...5
#>  1:  6234          0 10:27:22.818433 2101-10-09 FALSE
#>  2: -4561  -91401745 23:17:19.357471 2893-02-28  TRUE
#>  3:     0  448170383 14:21:20.932153 1263-12-30  TRUE
#>  4: -6602          0 18:14:29.799636 2899-04-24 FALSE
#>  5: -2112 -357756090 16:17:41.374953 1272-04-08 FALSE
#>  6:  1946 -560096854 00:08:24.591624 2730-03-30  TRUE
#>  7:  8641  700311762 23:23:52.226532 2845-07-16 FALSE
#>  8:  4741  222987195 19:34:46.593417 1509-04-18 FALSE
#>  9: -3566          0 10:29:55.632030 1553-09-10  TRUE
#> 10:  5965 -812109538 10:57:27.762808 1021-03-11  TRUE