data.table generator with randomized columns

data.table generator with randomized columns

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data.table_of(logical_(), date_()) %>% show_example()
#>          ...1  ...2  ...3  ...4       ...5  ...6       ...7       ...8
#> 1: 1701-11-11  TRUE FALSE FALSE 1708-11-27 FALSE 2874-08-17 2625-12-05
#> 2: 1512-01-23  TRUE  TRUE FALSE 2791-08-10  TRUE 2989-10-15 2650-09-16
#> 3: 2841-11-27  TRUE  TRUE FALSE 1920-11-15 FALSE 1773-10-23 2834-12-27
#> 4: 2257-11-05  TRUE FALSE  TRUE 2490-05-14  TRUE 1131-10-29 1276-07-11
#> 5: 1624-11-26  TRUE  TRUE  TRUE 2989-12-28 FALSE 2522-07-10 1939-11-10
#> 6: 1752-05-30  TRUE  TRUE FALSE 1399-08-22 FALSE 1605-07-20 1652-04-07
#> 7: 2525-01-10  TRUE FALSE FALSE 1945-11-13 FALSE 1740-06-01 2116-06-19
#> 8: 1907-09-23  TRUE  TRUE  TRUE 2001-03-24  TRUE 1732-11-26 2678-05-04
#> 9: 2999-10-07 FALSE FALSE  TRUE 2742-03-20 FALSE 1347-08-30 2801-11-26
#>          ...9
#> 1: 2445-03-18
#> 2: 1939-06-12
#> 3: 2537-05-09
#> 4: 2173-09-10
#> 5: 2658-03-08
#> 6: 2993-07-09
#> 7: 2751-05-03
#> 8: 1504-11-18
#> 9: 1262-10-05
data.table_of(any_atomic(), rows = 10L, cols = 5L) %>% show_example()
#>           ...1                ...2       ...3     ...4       ...5
#>  1:       PmU< 1969-07-14 17:30:55 -425389077     g*@4 -631553315
#>  2:   hRcx[7F) 1055-08-19 09:27:56  936798502     Q^jB -522857932
#>  3:       5FtQ 2869-02-03 08:28:02  565689121 'c#,E)Nv -859277503
#>  4:      aFl\\ 1780-10-31 05:36:11  938333939   0~[pjt -675407375
#>  5: C"\\rt2JOd 2324-06-12 15:33:17  840318006       ,Q   98721962
#>  6:  {>4o"1@Om  315-04-27 11:55:14  632697630   m0Mt9N   50827025
#>  7:       II/'  446-09-24 23:02:55 -718666352   \\aJI. -738584489
#>  8:  WANM{.w*b 2906-03-25 21:22:24          0       E)          0
#>  9:     :;xjzs 1515-09-15 19:39:45 -749641316   i*k#1a  -15831956
#> 10:          d 1513-02-17 17:09:35 -433293185  F"J>{Ts -431916328