Random data.table generator

Random data.table generator

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

data.table_of(logical_(), date_()) |> show_example()
#>     ...1  ...2  ...3  ...4       ...5       ...6       ...7       ...8  ...9
#> 1: FALSE FALSE  TRUE  TRUE 2419-06-09 2447-11-01 1238-08-16 2872-10-19  TRUE
#> 2:  TRUE FALSE  TRUE  TRUE 2073-01-10 2425-05-03 2203-05-13 1268-05-08  TRUE
#> 3:  TRUE FALSE FALSE  TRUE 1786-02-11 2920-10-16 1912-10-06 1675-04-29 FALSE
#> 4: FALSE FALSE FALSE FALSE 1127-07-06 2793-10-27 2398-04-05 2419-03-23 FALSE
#> 5: FALSE FALSE  TRUE  TRUE 2391-01-21 1265-07-16 2078-10-06 1079-07-12  TRUE
#> 6: FALSE  TRUE  TRUE  TRUE 1595-04-02 1729-05-11 2131-03-21 1842-10-25 FALSE
data.table_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>                ...1                ...2  ...3            ...4  ...5
#>  1: 08:55:33.142907  824-05-23 17:33:48  TRUE 10:10:13.189969  TRUE
#>  2: 15:30:09.937907 1955-05-26 06:20:54  TRUE 08:22:14.181413 FALSE
#>  3: 00:39:55.739079 2203-08-23 05:36:32  TRUE 07:03:19.313316 FALSE
#>  4: 12:23:12.070656 1571-08-05 00:59:54  TRUE 18:54:10.723920  TRUE
#>  5: 15:57:10.123704 1891-03-04 01:36:38  TRUE 15:57:15.158983 FALSE
#>  6: 19:44:59.399013 2597-07-21 19:42:58  TRUE 13:23:19.385620  TRUE
#>  7: 11:48:41.169668  979-09-25 04:18:30 FALSE 04:44:48.120977 FALSE
#>  8: 17:02:15.808672 2485-01-30 00:16:21  TRUE 04:53:28.408351 FALSE
#>  9: 19:31:38.043836 1169-07-04 10:05:40 FALSE 01:29:03.239530  TRUE
#> 10: 19:11:24.281487 2440-04-22 00:39:17 FALSE 22:28:52.520481  TRUE