data.table generator with randomized columns

data.table generator with randomized columns

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data.table_of(logical_(), date_()) %>% show_example()
#>          ...1       ...2  ...3  ...4  ...5       ...6       ...7       ...8
#> 1: 1321-09-15 1348-03-29  TRUE  TRUE  TRUE 2435-04-07 2075-06-27 1763-05-17
#> 2: 1577-05-10 1808-01-13  TRUE FALSE FALSE 1479-04-12 1787-08-03 1374-08-10
#> 3: 1076-08-17 2244-03-13  TRUE FALSE FALSE 2800-12-30 2930-04-24 1092-06-12
#> 4: 2809-12-31 1463-09-22  TRUE  TRUE  TRUE 1056-08-21 2449-03-25 2435-05-09
#> 5: 2101-11-14 2474-07-16 FALSE  TRUE FALSE 1926-10-29 2030-12-11 2972-11-16
data.table_of(any_atomic(), rows = 10L, cols = 5L) %>% show_example()
#>      ...1                ...2       ...3            ...4      ...5
#>  1: -3481  891-03-26 03:32:41  439787389 23:31:19.010951       XTc
#>  2:  7164 2581-10-24 14:45:01 -663899137 04:46:46.455316 'RNI,vzhg
#>  3: -8010 2920-02-18 00:29:09          0 21:29:19.684432     )`3[,
#>  4: -8632 2994-04-18 17:34:02 -679672316 10:41:34.812118    sYRzVE
#>  5:     0 2624-01-23 16:39:38  265832045 22:08:37.910751   cvfQwbB
#>  6:     0  324-05-19 15:06:02  960727292 11:21:30.219761         Q
#>  7:     0  352-04-02 10:36:11 -637565127 04:31:06.530517 ]m|W,N9\\
#>  8:  3932 2134-11-03 07:16:13 -880358712 16:30:23.159711   R5ou<~,
#>  9: -5970   68-10-26 09:52:00 -808915892 19:57:57.436241   `BVM|OJ
#> 10:  2931  650-11-20 22:47:37  890672815 01:49:24.828240 R%=;P8`]q