data.table generator with randomized columns

data.table generator with randomized columns

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data.table_of(logical_(), date_()) %>% show_example()
#>          ...1       ...2       ...3  ...4  ...5  ...6  ...7       ...8
#> 1: 2062-02-23 2879-11-08 1888-06-01  TRUE FALSE FALSE  TRUE 2749-10-17
#> 2: 2734-04-08 1405-06-24 1458-10-10  TRUE FALSE  TRUE  TRUE 2581-08-16
#> 3: 1379-06-18 1396-06-12 1257-05-06  TRUE  TRUE FALSE FALSE 1707-02-05
#> 4: 1250-10-14 1667-04-30 2872-03-03  TRUE  TRUE FALSE  TRUE 2453-04-27
#> 5: 1601-07-07 2858-11-08 2908-02-09 FALSE  TRUE  TRUE FALSE 2289-07-25
#> 6: 1639-07-21 1760-05-04 1889-04-24 FALSE  TRUE  TRUE  TRUE 2895-03-27
#> 7: 2144-05-12 1678-10-01 2850-08-19 FALSE FALSE FALSE  TRUE 1655-04-18
#> 8: 1542-06-19 2353-03-20 2388-07-05 FALSE FALSE FALSE FALSE 2584-08-06
#> 9: 1782-07-01 1749-12-17 2950-11-06 FALSE  TRUE FALSE  TRUE 2512-09-15
#>          ...9
#> 1: 1152-06-08
#> 2: 2659-06-18
#> 3: 2878-06-23
#> 4: 2281-07-13
#> 5: 1642-10-16
#> 6: 2331-07-05
#> 7: 1813-02-19
#> 8: 2569-03-29
#> 9: 2509-08-29
data.table_of(any_atomic(), rows = 10L, cols = 5L) %>% show_example()
#>                    ...1  ...2       ...3      ...4                ...5
#>  1: 2927-03-23 12:53:07  3370 -485971679  $vp(DvYy 2181-08-19 10:07:52
#>  2: 2251-08-24 14:51:11 -4215 -682373391         _ 1846-03-29 16:53:03
#>  3: 2088-07-06 06:22:46   169 -655147674   6!#^tIK 2667-03-31 19:32:10
#>  4: 1453-01-19 04:55:41  7782 -240459864    D$Zru$ 2189-09-26 03:55:40
#>  5: 1740-07-03 07:38:10   -21          0        G^ 1245-12-07 19:28:07
#>  6: 1948-12-07 21:26:00  7774  596492078        yN  974-03-11 00:01:39
#>  7: 1020-11-14 02:51:06   442  124292782    c8pnw/  128-12-06 22:08:54
#>  8: 2786-06-05 07:57:58     0 -538525549   r9J7Vdf 1122-11-27 10:16:43
#>  9: 2529-07-22 16:53:13 -2008 -742085282       yVo 1789-05-29 01:54:36
#> 10:  887-12-15 00:37:07  9248  569610885 ;giWyy2t% 1141-10-18 04:31:21