data.table generator with randomized columns

data.table generator with randomized columns

Usage

data.table_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data.tables will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data.table_of(logical_(), date_()) |> show_example()
#>          ...1  ...2  ...3       ...4       ...5       ...6  ...7  ...8
#> 1: 2003-07-18  TRUE  TRUE 1713-07-07 2301-06-03 1244-01-22 FALSE  TRUE
#> 2: 2294-04-10 FALSE  TRUE 1048-02-27 2207-06-24 2053-07-25 FALSE  TRUE
#> 3: 1549-12-17  TRUE  TRUE 2221-10-01 2479-07-05 2434-12-08  TRUE FALSE
#> 4: 1893-10-28  TRUE  TRUE 2570-05-28 1010-06-04 1183-02-03  TRUE FALSE
#> 5: 2937-09-08  TRUE FALSE 1056-06-19 1313-12-29 2243-06-30 FALSE FALSE
#> 6: 2671-08-12  TRUE FALSE 1849-03-27 1172-06-27 2320-08-06  TRUE FALSE
#> 7: 2261-03-03  TRUE  TRUE 1099-11-28 1619-08-19 2641-12-31 FALSE FALSE
data.table_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>           ...1      ...2  ...3       ...4  ...5
#>  1: 2737-12-03      9~f~ -8194 1690-08-11  6394
#>  2: 1965-01-23         <  8015 1103-12-23 -8394
#>  3: 1236-08-24    at0fDf  4613 1260-01-23  4095
#>  4: 1278-11-15     >&"EI -4763 1397-10-03  7246
#>  5: 1382-05-03    XsXP?T  9072 1489-11-24     0
#>  6: 1924-04-25 sR q$T&vH     0 1518-09-07 -6354
#>  7: 1353-12-07      7NPL  9836 1550-09-01     0
#>  8: 1341-05-01        c&  -907 2748-01-04 -5629
#>  9: 2652-07-30 YaArZuGb|  5018 1687-03-24 -5656
#> 10: 1514-06-22   fW:KhjY -5663 1284-09-25 -5908