Data frame generator with randomized columns

Data frame generator with randomized columns

Usage

data_frame_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data frames will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data_frame_of(logical_(), date_()) |> show_example()
#>    ...1
#> 1  TRUE
#> 2 FALSE
#> 3 FALSE
#> 4  TRUE
#> 5 FALSE
#> 6 FALSE
#> 7  TRUE
#> 8 FALSE
#> 9 FALSE
data_frame_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>          ...1      ...2       ...3    ...4      ...5
#> 1  -736716654         n 1101-12-28  F2z^k  X/8^$6 WY
#> 2  -413397759        }s 1140-05-30  *!`$a>   V]^'*oN
#> 3   315894002       vBy 1763-01-14     c9_        Q=
#> 4  -799855240       HVd 1013-12-04  l:~m[U         %
#> 5  -868354796 H!]r/{59m 1971-08-04 }{Vt7#h      P'Bw
#> 6   800081592    bs~Omm 1046-12-28  5L#9*2   E_~H gO
#> 7   354210921      1(UY 1255-10-09 Glj$,wq         t
#> 8  -849167864       Sw* 1651-08-09     <."    Qg<o*y
#> 9   482417823         @ 1299-07-19  bs~Omm      +"0Q
#> 10  160947305  "(~W[=bc 1361-11-11   N<ilP   "ye$?bR