Random data frame generator

Random data frame generator

Usage

data_frame_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data frames will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

data_frame_of(logical_(), date_()) |> show_example()
#>         ...1
#> 1 1384-04-15
#> 2 1557-03-19
#> 3 2025-12-11
data_frame_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>               ...1  ...2      ...3      ...4            ...5
#> 1  18:27:58.708517  7266    (4aU"r        =2 02:53:08.618905
#> 2  16:35:53.527501  8398        ko     Ng=I0 14:59:31.764003
#> 3  10:39:29.970708 -9513        O,      %=<7 14:48:28.637516
#> 4  08:29:12.130937 -4316 gPHaj@%7o     t}uoE 07:40:39.316619
#> 5  10:46:05.479978 -2502     2LK5! A[B*@FIT# 11:29:46.275260
#> 6  13:11:15.653806     0     mw<NJ     <Nn"@ 12:55:31.611274
#> 7  23:57:19.634493 -9299     j?>+[  k+Wv]Wq1 12:02:41.099788
#> 8  08:19:10.631063  7353 VK9Y](M0T        H` 21:47:39.083809
#> 9  05:59:26.346223 -1098  &zz4IxaQ         ^ 01:26:55.092483
#> 10 11:35:52.023036 -5360     %0~{8      ?#h' 21:18:08.626396