Data frame generator with randomized columns

Data frame generator with randomized columns

Usage

data_frame_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated data frames will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

data_frame_of(logical_(), date_()) |> show_example()
#>    ...1       ...2       ...3       ...4  ...5  ...6
#> 1  TRUE 2971-07-17 1816-02-05 2395-08-16  TRUE FALSE
#> 2  TRUE 1286-11-09 1677-06-22 2883-03-28  TRUE  TRUE
#> 3 FALSE 1576-10-23 2300-10-10 1527-04-12 FALSE FALSE
#> 4  TRUE 2275-03-27 2276-04-02 1439-08-01  TRUE  TRUE
#> 5 FALSE 2896-06-18 2603-06-17 1611-04-15 FALSE FALSE
#> 6 FALSE 2119-10-05 2611-08-11 2713-08-14  TRUE FALSE
#> 7 FALSE 1761-08-29 1869-01-11 1067-05-01  TRUE FALSE
data_frame_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#>               ...1   ...2     ...3      ...4                ...5
#> 1  23:12:37.210486 }<02P$        M        /G 1135-10-11 01:12:38
#> 2  09:06:44.029291     &V        )      9IP3  953-06-22 19:59:10
#> 3  18:10:52.516848      F JuCK/<N3    %B,nWc  732-08-24 02:43:23
#> 4  13:08:57.701152      D   -9%J|k wyIP!Q].4  504-04-27 02:23:04
#> 5  07:27:50.938049      )     /Hg]  1U aG43r 1397-04-09 18:51:43
#> 6  11:57:43.860523      L +_sA-PQG        gE  956-03-23 13:49:52
#> 7  07:13:26.917290      s      ~AD        h?  516-03-17 04:51:11
#> 8  12:23:08.851157   +A+6    XX%~-      pr*k  273-11-07 12:57:09
#> 9  22:33:46.826804  Bak\\        > 2@QDl!)8P 1000-04-04 00:53:02
#> 10 07:26:32.493275     .&       TB zAxALgR!2  756-10-31 19:58:28