Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

Value

A hedgehog.internal.gen generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 4 × 9
#>   ...1       ...2  ...3       ...4  ...5  ...6       ...7       ...8  ...9      
#>   <date>     <lgl> <date>     <lgl> <lgl> <date>     <date>     <lgl> <date>    
#> 1 1079-02-07 TRUE  2029-05-29 FALSE FALSE 2253-04-02 2967-01-08 TRUE  2796-06-25
#> 2 1306-03-23 TRUE  1909-08-03 TRUE  TRUE  2419-11-03 1949-11-06 FALSE 2513-02-26
#> 3 1770-08-19 FALSE 1164-07-04 TRUE  TRUE  2920-03-22 2146-09-26 TRUE  2766-02-20
#> 4 1032-09-06 TRUE  2304-07-22 FALSE TRUE  2589-10-26 2161-09-20 TRUE  1570-05-21
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>    ...1            ...2      ...3  ...4                ...5           
#>    <time>          <chr>     <lgl> <dttm>              <time>         
#>  1 04:18:39.679462 p:H|DTdT  FALSE 444-05-16 19:22:06  08:08:57.982293
#>  2 17:26:40.026751 X9d.a     TRUE  2169-04-25 10:43:25 15:05:20.145226
#>  3 04:39:20.221662 pd<At     FALSE 2925-03-17 23:07:35 23:00:35.666284
#>  4 04:21:41.030976 %8        TRUE  1268-12-02 00:38:49 03:05:32.414107
#>  5 13:21:21.499432 5         FALSE 874-10-26 15:51:44  10:59:29.952523
#>  6 02:18:48.803968 ,gEwi     FALSE 2155-04-25 02:39:14 00:27:09.781087
#>  7 16:27:51.454424 4WNj      FALSE 964-12-26 17:01:30  04:41:30.807353
#>  8 07:51:52.568221 o1C_>?h[H TRUE  2304-06-06 03:53:41 12:51:37.403869
#>  9 05:29:44.640299 I$;5,     FALSE 1558-11-10 07:05:49 10:03:07.283589
#> 10 10:14:36.462742 nw        FALSE 1562-04-29 12:13:18 21:28:18.144198