Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

Value

A hedgehog.internal.gen generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 7 × 10
#>   ...1       ...2       ...3       ...4       ...5  ...6       ...7  ...8 
#>   <date>     <date>     <date>     <date>     <lgl> <date>     <lgl> <lgl>
#> 1 1494-02-06 1592-08-14 1751-08-16 2234-12-02 TRUE  1895-01-25 FALSE TRUE 
#> 2 1230-03-24 2522-08-23 1399-07-27 2277-06-14 TRUE  1776-02-15 TRUE  TRUE 
#> 3 2682-04-26 1976-07-21 1526-02-16 2054-08-07 FALSE 2268-08-19 TRUE  FALSE
#> 4 1810-08-17 1971-06-27 2222-04-14 2921-11-19 TRUE  1526-05-29 TRUE  FALSE
#> 5 1191-10-22 2108-12-26 2362-04-01 1080-12-25 TRUE  2800-02-13 TRUE  FALSE
#> 6 1041-03-15 2927-01-02 2675-03-26 2536-08-22 FALSE 1601-05-03 TRUE  FALSE
#> 7 2150-07-29 1545-12-17 2072-03-12 2469-05-29 TRUE  1013-02-14 FALSE TRUE 
#> # … with 2 more variables: ...9 <date>, ...10 <date>
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>    ...1         ...2            ...3        ...4      ...5 
#>    <fct>        <time>          <fct>       <fct>     <lgl>
#>  1 "F\"5F\"}pL" 16:45:23.356157 "!\\g]/"    "sEc++d3" FALSE
#>  2 "7yW/"       22:50:39.151600 "eKFZg#yWI" "?[V"     TRUE 
#>  3 "We\\+vnnL"  05:58:14.719900 "FVE8&vuG"  "5$ "     TRUE 
#>  4 "o|Tf"       08:02:22.916166 "IGR+_0"    "X.5"     FALSE
#>  5 "@TNc"       17:18:09.109546 "`E.8P^QB"  "KHBl3n`" FALSE
#>  6 "DQ"         03:20:59.325368 "4Zk[>0Rbn" "AT"      FALSE
#>  7 "V7M@51=D"   04:51:07.044892 "Br"        "t^y$pM"  FALSE
#>  8 "hhP'6 <v"   04:56:43.086595 "emf\\b*r}" "c~"      TRUE 
#>  9 "k"          15:45:58.483360 "tj:3KtHBR" "C"       TRUE 
#> 10 " 'ZQ'Z"     08:01:26.011964 "4Zd,ieQx"  "I73x$ "  FALSE