Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 7 × 7
#>   ...1  ...2       ...3       ...4       ...5       ...6  ...7      
#>   <lgl> <date>     <date>     <date>     <date>     <lgl> <date>    
#> 1 FALSE 1229-06-11 2991-05-03 2967-08-02 2508-08-14 TRUE  1718-02-09
#> 2 FALSE 2422-12-26 2992-08-01 2131-10-27 2913-01-26 TRUE  2957-05-13
#> 3 TRUE  2287-10-11 2794-04-23 2671-08-02 1742-11-26 FALSE 1757-08-21
#> 4 FALSE 1983-11-14 2457-04-28 1700-04-07 2641-05-17 TRUE  2576-02-18
#> 5 TRUE  1736-11-05 2939-05-01 2383-01-17 2613-12-03 TRUE  1521-11-15
#> 6 FALSE 1310-10-16 1755-07-03 2672-04-05 1965-03-03 TRUE  2434-11-23
#> 7 FALSE 1895-01-22 1644-03-22 1474-02-03 1033-06-26 TRUE  2330-11-27
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>    ...1         ...2        ...3 ...4     ...5           
#>    <lgl>       <dbl>       <dbl> <chr>    <time>         
#>  1 TRUE           0  -179146948. ";5/HK|" 21:42:17.784201
#>  2 TRUE   551283573. -149703028. "YZL&"   14:44:07.833158
#>  3 TRUE  -607928436.  273419669. "ChzdT`" 02:01:17.304803
#>  4 FALSE -920823713.   73906519. "vZG"    09:39:32.611217
#>  5 TRUE   969443901.  983505067. "OXr"    20:08:59.896434
#>  6 FALSE -293055861.  188152274. "b}eS\\" 18:24:45.808457
#>  7 FALSE  106520799. -453901869. ";W]X'$" 03:35:36.192096
#>  8 TRUE   275092851.   -8807777. "KJN"    02:37:58.727552
#>  9 TRUE  -575524323. -701450474. "Jeq"    14:57:13.588480
#> 10 TRUE  -223501657. -968877342. "BL1"    18:48:07.498812