Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 7 × 6
#>   ...1       ...2  ...3       ...4       ...5  ...6      
#>   <date>     <lgl> <date>     <date>     <lgl> <date>    
#> 1 2399-07-03 FALSE 1454-09-15 1570-08-15 TRUE  1244-09-11
#> 2 2865-09-24 FALSE 1379-12-03 1692-12-15 TRUE  2174-02-22
#> 3 2226-02-17 FALSE 1595-08-03 2284-04-28 FALSE 1032-04-07
#> 4 1836-08-21 TRUE  1961-01-11 1213-05-28 TRUE  2951-06-15
#> 5 2916-11-20 TRUE  1452-11-05 2947-10-18 TRUE  1589-03-14
#> 6 1248-05-02 FALSE 1106-08-13 2768-07-26 FALSE 2823-05-22
#> 7 2728-03-26 FALSE 1324-03-16 2363-02-24 TRUE  1538-04-05
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>     ...1 ...2  ...3                ...4                 ...5
#>    <int> <lgl> <dttm>              <dttm>              <int>
#>  1  8161 FALSE 167-08-24 13:20:09  1536-05-15 11:05:33  6428
#>  2 -5221 FALSE 1872-11-06 13:23:59 51-09-19 20:08:23   -8753
#>  3  8895 TRUE  77-04-11 21:18:06   1757-01-10 13:53:10  2122
#>  4  7294 FALSE 1307-11-20 02:18:34 1656-04-14 16:00:48 -4587
#>  5  4552 TRUE  636-10-15 00:57:09  208-07-27 13:46:07  -4398
#>  6 -8130 FALSE 2648-08-11 06:44:14 1756-05-28 00:43:56  7837
#>  7 -7578 FALSE 2254-01-29 01:33:02 2410-04-08 11:45:31 -5265
#>  8  2655 FALSE 128-07-23 14:27:17  907-09-21 03:40:33      0
#>  9  4697 FALSE 433-08-01 17:38:23  2710-08-25 06:13:37  7804
#> 10 -2930 TRUE  2073-11-21 16:11:54 2751-11-18 15:01:37 -2087