Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

tibble_of(logical_(), date_()) %>% show_example()
#> # A tibble: 9 × 7
#>   ...1       ...2       ...3       ...4  ...5       ...6  ...7 
#>   <date>     <date>     <date>     <lgl> <date>     <lgl> <lgl>
#> 1 2839-09-27 2413-09-14 1200-11-18 TRUE  2719-10-16 TRUE  FALSE
#> 2 1046-10-07 2810-03-26 2556-01-27 TRUE  1689-11-04 TRUE  TRUE 
#> 3 1551-07-10 2335-06-25 1812-09-28 FALSE 1258-11-15 FALSE FALSE
#> 4 2720-06-19 2063-06-01 1884-07-07 FALSE 2246-02-20 FALSE FALSE
#> 5 1773-06-11 1806-02-25 2586-05-02 TRUE  1587-10-23 TRUE  FALSE
#> 6 1262-06-21 1820-06-05 2230-01-03 TRUE  1235-01-07 TRUE  FALSE
#> 7 1763-10-30 2387-01-15 2252-10-30 TRUE  1517-08-29 FALSE TRUE 
#> 8 1000-10-20 1508-05-15 1092-10-07 FALSE 1043-07-08 FALSE TRUE 
#> 9 2834-03-23 2565-01-17 2212-11-07 TRUE  1868-04-30 TRUE  FALSE
tibble_of(any_atomic(), rows = 10L, cols = 5L) %>% show_example()
#> # A tibble: 10 × 5
#>    ...1            ...2         ...3  ...4 ...5      
#>    <time>          <lgl>       <dbl> <int> <date>    
#>  1 10:57:59.114920 TRUE   931207399.  7928 1564-01-07
#>  2 19:50:57.523825 TRUE  -154801267.  3807 2119-11-27
#>  3 08:42:49.479204 TRUE   276299677. -2194 2879-05-30
#>  4 21:45:30.447733 FALSE  479886596.   658 1265-04-26
#>  5 23:02:58.607829 TRUE   520723031. -8251 1287-08-30
#>  6 06:39:21.169224 TRUE  -804049009.  1692 1745-02-12
#>  7 16:11:26.962589 FALSE  203941629. -2399 2477-09-11
#>  8 14:44:36.140362 FALSE  614050136.  3025 2284-10-06
#>  9 11:05:45.671745 TRUE   -31758024. -1375 2511-03-08
#> 10 10:20:30.016156 FALSE -190322214.   500 1074-01-22