Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 5 × 8
#>   ...1  ...2       ...3  ...4       ...5  ...6       ...7       ...8      
#>   <lgl> <date>     <lgl> <date>     <lgl> <date>     <date>     <date>    
#> 1 FALSE 2565-03-15 FALSE 2026-07-29 FALSE 1308-04-02 1013-10-15 1717-08-11
#> 2 FALSE 2018-12-22 FALSE 2272-05-21 FALSE 1922-12-22 1622-02-26 1077-05-18
#> 3 TRUE  1903-09-02 TRUE  1616-04-26 TRUE  2822-02-20 1699-07-10 1642-04-04
#> 4 FALSE 1046-04-20 FALSE 1947-12-19 FALSE 2095-04-04 2012-07-29 2714-10-14
#> 5 FALSE 2236-09-21 FALSE 1953-05-17 FALSE 1870-01-16 2750-09-01 2998-03-09
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>    ...1        ...2       ...3                ...4                ...5 
#>    <fct>       <date>     <dttm>              <dttm>              <lgl>
#>  1 " "         2586-06-09 2909-09-08 16:53:12 90-12-06 16:36:37   FALSE
#>  2 "m"         1561-06-14 2660-10-11 17:12:45 1459-06-07 08:40:12 TRUE 
#>  3 "tCT98"     1725-10-22 27-06-12 01:08:12   1979-04-05 13:45:12 FALSE
#>  4 ",L;rfd"    1940-06-07 154-08-13 13:17:31  1264-01-03 09:28:28 FALSE
#>  5 "H\"sRo[S"  2870-06-27 2561-10-25 11:41:30 2527-01-12 05:20:42 FALSE
#>  6 "Ym1<X194P" 1045-11-24 329-09-11 23:38:19  595-01-02 07:21:07  TRUE 
#>  7 " w"        1319-01-15 1255-02-02 16:23:47 1410-03-22 15:16:56 FALSE
#>  8 "Zu/2%e]}"  2968-07-07 2103-03-10 14:54:09 307-09-29 22:22:12  TRUE 
#>  9 "a"         1435-02-15 1941-09-29 14:06:04 2129-09-25 22:41:47 TRUE 
#> 10 "'7^@^O,"   2021-04-08 2326-12-31 11:36:35 895-06-11 14:15:15  FALSE