Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 9 × 1
#>   ...1      
#>   <date>    
#> 1 2539-04-06
#> 2 2365-02-15
#> 3 2855-05-30
#> 4 1325-11-26
#> 5 2803-09-10
#> 6 2236-05-03
#> 7 2033-02-10
#> 8 2493-03-21
#> 9 1071-10-24
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>           ...1 ...2       ...3            ...4            ...5 
#>          <dbl> <date>     <time>          <time>          <lgl>
#>  1  239874369. 1649-09-08 13:48:28.474291 23:18:27.841802 TRUE 
#>  2  630694874. 2640-08-17 18:22:28.377017 09:06:07.013665 FALSE
#>  3  984971650. 1500-09-07 12:18:12.752006 21:25:14.641444 TRUE 
#>  4 -384132444. 2129-10-21 19:47:15.686638 09:43:38.480967 FALSE
#>  5 -843413780. 1385-08-16 17:38:14.360010 23:11:34.715109 TRUE 
#>  6 -365322415. 1803-05-11 13:09:05.590960 18:27:14.064134 FALSE
#>  7  -64795796. 2501-03-23 10:08:38.717927 07:40:04.358795 TRUE 
#>  8 -156607223. 2874-03-08 16:34:06.774256 01:45:19.558883 TRUE 
#>  9   56591524. 2443-04-12 22:46:56.515613 18:42:18.335808 TRUE 
#> 10          0  2823-06-05 12:12:53.986574 18:12:37.872490 FALSE