Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

Value

A hedgehog.internal.gen generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 1 × 5
#>   ...1  ...2       ...3       ...4  ...5 
#>   <lgl> <date>     <date>     <lgl> <lgl>
#> 1 FALSE 2665-06-06 2636-05-30 TRUE  FALSE
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>    ...1                ...2       ...3        ...4            ...5           
#>    <dttm>              <date>     <chr>       <time>          <time>         
#>  1 2292-06-02 18:10:53 1532-11-24 "g@AO"      23:41:43.315466 15:26:38.904681
#>  2 1893-04-30 04:52:49 1822-04-24 "KGnhK:EeU" 13:46:28.017537 15:13:30.659600
#>  3 388-03-27 21:29:23  2695-08-12 "*Mrv(}}z"  00:29:47.309772 16:17:05.222963
#>  4 1881-12-12 14:13:30 2854-12-09 "|;"        05:55:30.052693 00:19:09.834589
#>  5 2454-09-15 05:40:30 1228-06-09 "2#xs"      18:16:49.849679 02:48:09.461689
#>  6 1568-07-02 09:37:28 2437-04-24 "\"k^maP"   17:19:13.243519 13:33:08.878275
#>  7 637-09-09 23:01:31  1676-08-11 "-"         22:28:28.668201 15:35:25.884598
#>  8 1666-01-03 18:53:58 1194-05-08 "13"        11:59:51.950138 21:16:31.127872
#>  9 853-06-18 12:11:36  2958-01-02 "OyrKNW:f"  21:48:33.324146 05:17:08.537983
#> 10 973-05-05 17:14:19  2633-07-29 "I*7f4h9bw" 04:39:32.783790 13:33:00.429072