Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 6 × 4
#>   ...1       ...2       ...3  ...4 
#>   <date>     <date>     <lgl> <lgl>
#> 1 2277-07-06 1440-09-06 TRUE  TRUE 
#> 2 2947-10-27 2632-07-07 FALSE TRUE 
#> 3 2075-02-07 2770-07-04 TRUE  FALSE
#> 4 2015-12-25 2036-12-11 FALSE FALSE
#> 5 1691-05-09 2007-07-14 TRUE  TRUE 
#> 6 2450-01-11 1492-04-21 FALSE FALSE
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>     ...1 ...2            ...3        ...4  ...5      
#>    <int> <time>          <fct>       <lgl> <date>    
#>  1 -6361 15:04:08.353387 "jnb\"4}t"  FALSE 1254-08-04
#>  2 -5542 11:28:17.793722 "9}(6["     FALSE 1469-01-05
#>  3 -5056 17:38:19.658170 "fn9an+X{"  FALSE 1296-08-12
#>  4  7535 21:27:27.540573 "['TI_"     TRUE  1211-11-24
#>  5  2461 13:35:44.140210 "43oSaxs\"" TRUE  1703-07-23
#>  6 -8813 11:56:45.892428 "Pt"        FALSE 1823-06-23
#>  7 -5031 12:23:05.340857 "Sj'OWkYt-" FALSE 1527-04-03
#>  8 -6445 11:05:36.096470 "&c]UU&H'"  FALSE 1747-10-13
#>  9  7801 19:47:28.333155 "wS=~)"     FALSE 2494-01-18
#> 10  -690 04:35:12.379708 "ck"        FALSE 1702-01-13