Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

Value

A hedgehog.internal.gen generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 2 × 10
#>   ...1       ...2       ...3  ...4       ...5       ...6  ...7  ...8       ...9 
#>   <date>     <date>     <lgl> <date>     <date>     <lgl> <lgl> <date>     <lgl>
#> 1 2978-10-20 2385-01-17 FALSE 2547-05-25 2253-01-12 FALSE TRUE  2604-10-14 TRUE 
#> 2 1360-01-20 1268-05-25 TRUE  1260-06-12 1326-03-15 TRUE  FALSE 1831-12-19 FALSE
#> # … with 1 more variable: ...10 <date>
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>           ...1 ...2       ...3  ...4            ...5           
#>          <dbl> <date>     <lgl> <time>          <time>         
#>  1  883879828. 2473-05-05 TRUE  23:58:55.121162 04:00:47.405676
#>  2  352740188. 1382-09-21 TRUE  17:30:34.226029 09:25:25.917120
#>  3          0  2297-12-04 FALSE 09:59:47.424464 22:03:30.602175
#>  4  716725077. 1497-06-13 TRUE  10:14:29.587320 09:13:20.666926
#>  5  779210213. 1566-09-02 FALSE 05:36:14.279119 11:45:39.189056
#>  6  413542248. 1980-02-25 TRUE  16:48:44.542116 10:32:01.492795
#>  7  197367499. 1569-01-28 TRUE  06:40:06.712083 20:56:51.230266
#>  8  935417774. 2690-03-26 TRUE  19:04:42.799597 09:07:00.751617
#>  9  545198723. 1551-12-16 FALSE 18:46:51.656448 04:34:51.122616
#> 10 -573788263. 2989-07-22 FALSE 04:56:14.993610 16:28:41.115338