Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

Value

A hedgehog.internal.gen generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 3 x 10
#>   ...1       ...2  ...3       ...4  ...5       ...6       ...7       ...8      
#>   <date>     <lgl> <date>     <lgl> <date>     <date>     <date>     <date>    
#> 1 1878-08-25 FALSE 2076-03-11 FALSE 1785-07-02 1887-01-05 2404-11-08 1344-05-16
#> 2 2513-08-11 FALSE 1313-06-21 TRUE  2680-04-02 1062-11-21 1600-06-16 2518-11-08
#> 3 2740-05-24 TRUE  1951-12-08 TRUE  2705-03-28 2440-02-22 1022-10-28 1612-10-26
#> # … with 2 more variables: ...9 <date>, ...10 <date>
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 x 5
#>    ...1     ...2       ...3      ...4        ...5
#>    <chr>    <date>     <fct>    <int>       <dbl>
#>  1 >#r      2207-08-03 E]-^I^     531 -206845664.
#>  2 5%       1848-11-08 'x        -191  880456653.
#>  3 ZIA      1515-10-30 $V)f2'&/ -1587  377723951.
#>  4 >;iW     1912-02-01 $#0N     -2645 -781286725.
#>  5 N        1448-07-17 /H2~     -1942 -589876060.
#>  6 $        1651-09-01 M)7       9266  379159861.
#>  7 Gi       1264-03-31 T':'y9   -9240  752080618.
#>  8 {?ygA    2791-04-12 c4wN;    -9252 -524970485.
#>  9 f^_#O5va 2351-05-26 =YyN      4233          0 
#> 10 CC=8sri  1732-12-13 }         9864  333832284.