Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10. To produce empty tibbles set rows = 0L or a range like rows = c(0L, 10L).

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10. To produce empty tibbles set cols = 0L or a range like cols = c(0L, 10L).

Value

A quickcheck_generator object.

Examples

tibble_of(logical_(), date_()) %>% show_example()
#> # A tibble: 9 × 2
#>   ...1       ...2      
#>   <date>     <date>    
#> 1 1276-04-19 1292-08-31
#> 2 2150-09-10 2054-07-25
#> 3 1280-07-17 2329-07-23
#> 4 2487-05-25 2562-02-15
#> 5 1317-03-06 1433-07-20
#> 6 1434-09-16 1458-05-05
#> 7 1865-08-02 1405-05-08
#> 8 2566-02-14 1041-07-09
#> 9 2736-04-21 1107-12-31
tibble_of(any_atomic(), rows = 10L, cols = 5L) %>% show_example()
#> # A tibble: 10 × 5
#>    ...1  ...2       ...3                ...4        ...5           
#>    <lgl> <fct>      <dttm>              <fct>       <time>         
#>  1 TRUE  ":^_Y^"    1685-09-06 07:39:43 "e%B~"      15:11:15.011594
#>  2 TRUE  "Y"        2189-08-12 18:31:34 "B5A"       10:55:03.708789
#>  3 TRUE  "@ORx1&g"  2716-12-09 16:05:22 "t\\"       18:03:49.885589
#>  4 TRUE  "1^8"      479-11-28 12:59:22  "uRM!,NA{"  22:25:44.803414
#>  5 TRUE  ".51US_"   2966-03-10 10:08:05 "86nbj#mjO" 07:24:30.958105
#>  6 FALSE "*R5>[UD"  1172-05-20 22:03:03 "(G:.Z3"    23:34:06.263610
#>  7 FALSE "x?\"UrJ"  2829-08-24 01:25:16 "E+hBRXm-"  14:27:51.149841
#>  8 FALSE "wPU"      1601-03-25 02:44:59 "0i&x,"     20:04:50.304258
#>  9 TRUE  "<Pyn"     695-07-02 23:31:00  "[9."       21:16:33.929052
#> 10 TRUE  "rW-XjtbR" 1511-02-03 18:30:51 "(e5ipA8^"  09:31:51.167804