Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

Value

A hedgehog.internal.gen generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 1 × 8
#>   ...1  ...2  ...3  ...4  ...5  ...6  ...7       ...8 
#>   <lgl> <lgl> <lgl> <lgl> <lgl> <lgl> <date>     <lgl>
#> 1 TRUE  TRUE  TRUE  FALSE FALSE TRUE  1972-11-19 FALSE
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>    ...1        ...2            ...3       ...4        ...5           
#>    <chr>       <time>          <date>     <fct>       <time>         
#>  1 "rKHT\\7"   18:47:32.060348 1483-10-28 "`-I"       16:47:09.556846
#>  2 "'ql"       07:41:40.362775 2202-12-16 "<&Lt`K@9"  22:28:31.462179
#>  3 "~_-Yn"     13:51:53.096361 2226-08-18 "m%N,"      23:12:53.914668
#>  4 "WIt$aR TH" 00:37:54.158688 1526-07-27 "f?"        07:40:04.280120
#>  5 "&2oq;"     06:36:42.806798 1098-06-03 "G$mH:u"    06:28:12.799032
#>  6 "d~"        16:10:38.103472 2755-07-04 "/uFTJ"     15:16:37.986651
#>  7 "h<6R{0"    13:11:01.073145 1368-04-25 "8@v X*1B_" 02:40:56.191926
#>  8 "6va"       14:29:48.171086 1323-07-18 "gY;jw@s"   21:32:10.060313
#>  9 "L/sU"      08:27:00.584873 2208-03-21 "G8f^ \"A"  04:22:56.993737
#> 10 "NRq17,"    22:01:33.501208 2523-07-08 "ai5yl"     18:25:30.118969