Random tibble generator

Random tibble generator

Usage

tibble_of(..., rows = c(1L, 10L), cols = c(1L, 10L))

Arguments

...

A set of unnamed generators. The generated tibbles will be built with random combinations of these generators.

rows

Number of rows of the generated data frame. If rows is a single number all data frames will have this number of rows. If rows is a numeric vector of length 2 it will produce data frames with rows between a minimum and maximum, inclusive. For example rows = c(1L, 10L) would produce data frames with rows between 1 and 10.

cols

Number of columns of the generated data frame. If cols is a single number all data frames will have this number of columns. If cols is a numeric vector of length 2 it will produce data frames with columns between a minimum and maximum, inclusive. For example cols = c(1L, 10L) would produce data frames with columns between 1 and 10.

Value

A hedgehog.internal.gen generator

Examples

tibble_of(logical_(), date_()) |> show_example()
#> # A tibble: 6 × 1
#>   ...1      
#>   <date>    
#> 1 2873-12-17
#> 2 1042-12-18
#> 3 1134-09-28
#> 4 2785-12-17
#> 5 2581-11-21
#> 6 1816-09-28
tibble_of(any_atomic(), rows = 10L, cols = 5L) |> show_example()
#> # A tibble: 10 × 5
#>    ...1            ...2                ...3            ...4        ...5      
#>    <time>          <dttm>              <time>          <fct>       <date>    
#>  1 10:06:42.704128 1161-02-04 00:16:43 04:27:05.508865 ";Vy0:&\\"  1892-12-26
#>  2 02:53:14.206478 2092-10-02 14:37:36 08:07:24.227922 "xTKY#Ns"   2638-09-17
#>  3 13:45:18.848486 2360-04-12 07:24:34 08:38:14.561869 "admbl"     1064-11-12
#>  4 17:44:25.584318 1820-10-17 10:29:36 23:08:56.070552 "_J'$3qQ"   2054-01-27
#>  5 11:44:10.563746 2627-02-05 20:45:23 15:15:49.000841 "8u8{"      2841-06-13
#>  6 13:54:35.307665 748-06-22 18:09:17  01:58:35.400132 "L9*[~sR%"  2854-03-20
#>  7 23:28:49.402068 590-12-04 05:04:16  17:34:12.183017 "z\""       1306-08-30
#>  8 01:15:17.566057 1088-05-11 20:48:01 16:24:08.931874 ">1"        1051-08-29
#>  9 22:36:52.155277 1501-07-05 07:09:00 11:54:50.821280 " PU?,So[/" 2740-07-26
#> 10 23:10:32.399492 714-04-08 12:44:45  18:50:40.817490 "WI/2h"     1337-03-28